Update MenuBar component
