+++ cloud_dashboard/src/service/utility.ts
@@ -1,3 +1,4 @@
+import { OAUTH2_CLIENT_SECRET, ROUTE_URL } from "constant";
 import L from "leaflet";
 import CloudContenxtItem from "model/CloudContenxtItem";
 import CloudContext from "model/CloudContext";
@@ -355,3 +356,88 @@ export const convertCloudContenxtItemList = (
   }
   return Object.values(hash);
 }
+
+/**
+ * Update the access token using Code Grant.
+ *
+ * @param clientId Client ID.
+ * @@aram refreshToken Refresh token for refresh access token.
+ */
+const refreshTokenByCodeGrant = async (clientId: string, refreshToken: string) => {
+  const formData = new FormData();
+  formData.append('grant_type', 'refresh_token');
+  formData.append('client_id', clientId);
+  formData.append('client_secret', OAUTH2_CLIENT_SECRET);
+  formData.append('refresh_token', refreshToken);
+
+  const response = await fetch(`/oauth/token`, {
+    method: 'POST',
+    body: formData
+  });
+
+  if (!response.ok) {
+    throw new Error('Refresh failed');
+  }
+  const response_json = await response.json();
+  if ('access_token' in response_json) {
+    // read tokens
+    const accessToken = response_json['access_token'];
+    const refreshToken2 = response_json['refresh_token'];
+    const expiresDatetime = (new Date()).getTime() + response_json['expires_in'] * 1000;
+
+    // save tokens to Localstrage
+    window.localStorage.setItem('accessToken', accessToken);
+    window.localStorage.setItem('refreshToken', refreshToken2);
+    window.localStorage.setItem('expiresDatetime', `${expiresDatetime}`);
+  } else {
+    throw new Error('Refresh failed');
+  }
+};
+
+/**
+ * Check the status of the access token,
+ * and if there is a problem, go to the login screen.
+ */
+export const checkAndRefreshToken = async () => {
+  // If you don't have the access token, redirect route URL.
+  const accessToken = window.localStorage.getItem('accessToken');
+  const expiresDatetime = window.localStorage.getItem('expiresDatetime');
+  if (accessToken === null || expiresDatetime == null) {
+    window.location.href = ROUTE_URL;
+    return;
+  }
+
+  // If the information required to update the token cannot be loaded,
+  // redirect route URL.
+  console.group('Authorization Code Grant');
+  const res1 = await fetch('/clouds/cloud_dashboard/config/client_id');
+  const refreshToken = window.localStorage.getItem('refreshToken');
+  if (!res1.ok || refreshToken === null) {
+    console.log('Client ID : No');
+    console.error('Authorization failed.');
+    console.groupEnd();
+    window.location.href = ROUTE_URL;
+    return;
+  }
+  const clientId = (await res1.json()).id;
+  console.log('Client ID : Yes');
+
+  // If the access token has expired, update it.
+  const now = (new Date()).getTime();
+  if (now <= parseInt(expiresDatetime, 10)) {
+    console.log('Token expired : No');
+    console.groupEnd();
+    return;
+  }
+  console.log('Token expired : Yes');
+
+  refreshTokenByCodeGrant(clientId, refreshToken).then(() => {
+    console.log('Access token : Yes');
+    console.groupEnd();
+  }).catch(() => {
+    console.log('Access token : No');
+    console.error('Authorization failed.');
+    console.groupEnd();
+    window.location.href = ROUTE_URL;
+  });
+};
