+++ js/main.chunk.js
@@ -1,2 +1,2 @@
-(this.webpackJsonpcloud_dashboard=this.webpackJsonpcloud_dashboard||[]).push([[0],{56:function(e,a,t){"use strict";t.r(a);var n=t(1),r=t.n(n),l=t(20),c=t.n(l),o=t(6),i=t(11),s=t(5),u="cloud_dashboard",m="/clouds/dashboard",d=[{cloudServiceProvider:"aws_cloud",labelName:"Instance",entityName:"instance",entityColumn:[{labelName:"Name",name:"name",type:"default"},{labelName:"Public IP",name:"public_ip",type:"default"},{labelName:"Instance State",name:"instance_state",type:"default"},{labelName:"Instance Type",name:"instance_type",type:"default"},{labelName:"Availability Zone",name:"availability_zone",type:"default"},{labelName:"Cost",name:"cost",type:"cost"},{labelName:"Created",name:"created",type:"datetime"}]},{cloudServiceProvider:"aws_cloud",labelName:"Image",entityName:"image",entityColumn:[{labelName:"Name",name:"name",type:"default"},{labelName:"AMI Name",name:"ami_name",type:"default"},{labelName:"AMI ID",name:"image_id",type:"default"},{labelName:"Status",name:"status",type:"default"},{labelName:"Source",name:"source",type:"default"},{labelName:"Root Device Type",name:"root_device_type",type:"default"},{labelName:"Visibility",name:"visibility",type:"boolean",value:["Public","Private"]},{labelName:"Created",name:"created",type:"datetime"}]},{cloudServiceProvider:"aws_cloud",labelName:"Security Group",entityName:"security_group",entityColumn:[{labelName:"Name",name:"name",type:"default"},{labelName:"Group ID",name:"group_id",type:"default"},{labelName:"VPC",name:"vpc_id",type:"join",info:{entityTypeId:"aws_cloud_vpc",keyColumn:"vpc_id",valueColumn:"name"}}]},{cloudServiceProvider:"aws_cloud",labelName:"Elastic IP",entityName:"elastic_ip",entityColumn:[{labelName:"Name",name:"name",type:"default"},{labelName:"Allocation ID",name:"allocation_id",type:"default"},{labelName:"Type",name:"elastic_ip_type",type:"default"},{labelName:"Elastic IP",name:"public_ip",type:"default"},{labelName:"Instance ID",name:"instance_id",type:"join",info:{entityTypeId:"aws_cloud_instance",keyColumn:"instance_id",valueColumn:"name"}},{labelName:"Private IP Address",name:"private_ip_address",type:"default"},{labelName:"Scope",name:"scope",type:"default"}]},{cloudServiceProvider:"aws_cloud",labelName:"Key Pair",entityName:"key_pair",entityColumn:[{labelName:"Key Pair Name",name:"key_pair_name",type:"default"},{labelName:"Key Fingerprint",name:"key_fingerprint",type:"default"}]},{cloudServiceProvider:"aws_cloud",labelName:"Volume",entityName:"volume",entityColumn:[{labelName:"Name",name:"name",type:"default"},{labelName:"Volume ID",name:"volume_id",type:"default"},{labelName:"IOPS",name:"iops",type:"default"},{labelName:"Size",name:"size",type:"default"},{labelName:"Availability Zone",name:"availability_zone",type:"default"},{labelName:"Volume Type",name:"volume_type",type:"default"},{labelName:"Attachment Information",name:"attachment_information",type:"join",info:{entityTypeId:"aws_cloud_instance",keyColumn:"instance_id",valueColumn:"name"}},{labelName:"State",name:"state",type:"default"},{labelName:"Created",name:"created",type:"datetime"}]},{cloudServiceProvider:"aws_cloud",labelName:"Snapshot",entityName:"snapshot",entityColumn:[{labelName:"Name",name:"name",type:"default"},{labelName:"Snapshot ID",name:"snapshot_id",type:"default"},{labelName:"Encrypted",name:"encrypted",type:"default"},{labelName:"Size",name:"size",type:"default"},{labelName:"Status",name:"status",type:"default"},{labelName:"Created",name:"created",type:"datetime"}]},{cloudServiceProvider:"aws_cloud",labelName:"Network Interface",entityName:"network_interface",entityColumn:[{labelName:"Name",name:"name",type:"default"},{labelName:"Subnet",name:"subnet_id",type:"join",info:{entityTypeId:"aws_cloud_subnet",keyColumn:"subnet_id",valueColumn:"name"}},{labelName:"VPC",name:"vpc_id",type:"join",info:{entityTypeId:"aws_cloud_vpc",keyColumn:"vpc_id",valueColumn:"name"}},{labelName:"Status",name:"status",type:"default"},{labelName:"Security Groups",name:"security_groups",type:"default"},{labelName:"Primary Private IP",name:"primary_private_ip",type:"default"},{labelName:"Secondary Private IPs",name:"secondary_private_ips",type:"default"}]},{cloudServiceProvider:"aws_cloud",labelName:"VPC",entityName:"vpc",entityColumn:[{labelName:"Name",name:"name",type:"default"},{labelName:"VPC",name:"vpc_id",type:"default"},{labelName:"State",name:"state",type:"default"},{labelName:"IPv4 CIDR",name:"cidr_block",type:"default"},{labelName:"IPv6 CIDR",name:"ipv6_cidr_blocks",type:"array"},{labelName:"Created",name:"created",type:"datetime"}]},{cloudServiceProvider:"aws_cloud",labelName:"Subnet",entityName:"subnet",entityColumn:[{labelName:"Name",name:"name",type:"default"},{labelName:"Subnet ID",name:"subnet_id",type:"default"},{labelName:"VPC",name:"vpc_id",type:"join",info:{entityTypeId:"aws_cloud_vpc",keyColumn:"vpc_id",valueColumn:"name"}},{labelName:"IPv4 CIDR",name:"cidr_block",type:"default"},{labelName:"State",name:"state",type:"default"},{labelName:"Region Name",name:"region_name",type:"default"},{labelName:"Zone Name",name:"zone_name",type:"default"},{labelName:"Network Border Group",name:"network_border_group",type:"default"},{labelName:"Zone Type",name:"zone_type",type:"default"},{labelName:"Parent Zone Type",name:"parent_zone_type",type:"default"},{labelName:"Created",name:"created",type:"datetime"}]},{cloudServiceProvider:"aws_cloud",labelName:"VPC Peering Connection",entityName:"vpc_peering_connection",entityColumn:[{labelName:"Name",name:"name",type:"default"},{labelName:"VPC Peering Connection ID",name:"vpc_peering_connection_id",type:"default"},{labelName:"Status",name:"status_code",type:"default"},{labelName:"Requester VPC",name:"requester_vpc_id",type:"join",info:{entityTypeId:"aws_cloud_vpc",keyColumn:"vpc_id",valueColumn:"name"}},{labelName:"Accepter VPC",name:"accepter_vpc_id",type:"join",info:{entityTypeId:"aws_cloud_vpc",keyColumn:"vpc_id",valueColumn:"name"}},{labelName:"Requester CIDR Blocks",name:"requester_cidr_block",type:"default"},{labelName:"Accepter CIDR Blocks",name:"accepter_cidr_block",type:"default"},{labelName:"Requester AWS Account",name:"requester_account_id",type:"default"},{labelName:"Accepter AWS Account",name:"accepter_account_id",type:"default"},{labelName:"Created",name:"created",type:"datetime"}]},{cloudServiceProvider:"aws_cloud",labelName:"Internet Gateway",entityName:"internet_gateway",entityColumn:[{labelName:"Name",name:"name",type:"default"},{labelName:"Internet Gateway ID",name:"internet_gateway_id",type:"default"},{labelName:"State",name:"state",type:"default"},{labelName:"VPC ID",name:"vpc_id",type:"default"},{labelName:"Created",name:"created",type:"default"}]},{cloudServiceProvider:"aws_cloud",labelName:"Career Gateway",entityName:"career_gateway",entityColumn:[{labelName:"Name",name:"name",type:"default"},{labelName:"Carrier Gateway ID",name:"carrier_gateway_id",type:"default"},{labelName:"State",name:"state",type:"default"},{labelName:"VPC ID",name:"vpc_id",type:"default"},{labelName:"Created",name:"created",type:"default"}]},{cloudServiceProvider:"aws_cloud",labelName:"Transit Gateway",entityName:"transit_gateway",entityColumn:[{labelName:"Name",name:"name",type:"default"},{labelName:"Transit Gateway ID",name:"transit_gateway_id",type:"default"},{labelName:"State",name:"state",type:"default"},{labelName:"Amazon Account ID",name:"account_id",type:"default"},{labelName:"Created",name:"created",type:"default"}]}],p=[{cloudServiceProvider:"k8s",labelName:"Node",entityName:"node",entityColumn:[{labelName:"Name",name:"name",type:"default"},{labelName:"State",name:"status",type:"default"},{labelName:"CPU (Request)",name:"cpu_request",type:"default"},{labelName:"CPU (Limit)",name:"cpu_limit",type:"default"},{labelName:"CPU (Usage)",name:"cpu_usage",type:"default"},{labelName:"Memory (Request)",name:"memory_request",type:"memory"},{labelName:"Memory (Limit)",name:"memory_limit",type:"memory"},{labelName:"Memory (Usage)",name:"memory_usage",type:"memory"},{labelName:"Pods (Applcation)",name:"pods_allocation",type:"default"},{labelName:"Created",name:"created",type:"datetime"}]},{cloudServiceProvider:"k8s",labelName:"Namespace",entityName:"namespace",entityColumn:[{labelName:"Name",name:"name",type:"default"},{labelName:"State",name:"status",type:"default"},{labelName:"Created",name:"created",type:"datetime"}]},{cloudServiceProvider:"k8s",labelName:"Deployment",entityName:"deployment",entityColumn:[{labelName:"Namespace",name:"namespace",type:"default"},{labelName:"Name",name:"name",type:"default"},{labelName:"Created",name:"created",type:"datetime"}]},{cloudServiceProvider:"k8s",labelName:"Pod",entityName:"pod",entityColumn:[{labelName:"Namespace",name:"namespace",type:"default"},{labelName:"Name",name:"name",type:"default"},{labelName:"Node",name:"node_name",type:"default"},{labelName:"State",name:"status",type:"default"},{labelName:"Restarts",name:"restarts",type:"default"},{labelName:"CPU (Usage)",name:"cpu_usage",type:"default"},{labelName:"Memory (Usage)",name:"memory_usage",type:"default"},{labelName:"Created",name:"created",type:"datetime"}]},{cloudServiceProvider:"k8s",labelName:"ReplicaSet",entityName:"replica_set",entityColumn:[{labelName:"Namespace",name:"namespace",type:"default"},{labelName:"Name",name:"name",type:"default"},{labelName:"Replica",name:"replicas",type:"default"},{labelName:"Created",name:"created",type:"datetime"}]},{cloudServiceProvider:"k8s",labelName:"CronJob",entityName:"cron_job",entityColumn:[{labelName:"Namespace",name:"namespace",type:"default"},{labelName:"Name",name:"name",type:"default"},{labelName:"Schedule",name:"schedule",type:"default"},{labelName:"Suspend",name:"suspend",type:"default"},{labelName:"Active",name:"active",type:"default"},{labelName:"Created",name:"created",type:"datetime"}]},{cloudServiceProvider:"k8s",labelName:"Job",entityName:"job",entityColumn:[{labelName:"Namespace",name:"namespace",type:"default"},{labelName:"Name",name:"name",type:"default"},{labelName:"Image",name:"image",type:"default"},{labelName:"Created",name:"created",type:"datetime"}]},{cloudServiceProvider:"k8s",labelName:"Service",entityName:"service",entityColumn:[{labelName:"Namespace",name:"namespace",type:"default"},{labelName:"Name",name:"name",type:"default"},{labelName:"Cluster IP",name:"cluster_ip",type:"default"},{labelName:"Created",name:"created",type:"datetime"}]},{cloudServiceProvider:"k8s",labelName:"Network Policy",entityName:"network_policy",entityColumn:[{labelName:"Namespace",name:"namespace",type:"default"},{labelName:"Name",name:"name",type:"default"},{labelName:"Created",name:"created",type:"datetime"}]},{cloudServiceProvider:"k8s",labelName:"Resource Quotas",entityName:"resource_quota",entityColumn:[{labelName:"Namespace",name:"namespace",type:"default"},{labelName:"Name",name:"name",type:"default"},{labelName:"Created",name:"created",type:"datetime"}]},{cloudServiceProvider:"k8s",labelName:"LimitRange",entityName:"limit_range",entityColumn:[{labelName:"Namespace",name:"namespace",type:"default"},{labelName:"Name",name:"name",type:"default"},{labelName:"Created",name:"created",type:"datetime"}]},{cloudServiceProvider:"k8s",labelName:"Priority Class",entityName:"priority_class",entityColumn:[{labelName:"Name",name:"name",type:"default"},{labelName:"Value",name:"value",type:"default"},{labelName:"Global Default",name:"global_default",type:"boolean",value:["TRUE","FALSE"]},{labelName:"Description",name:"description",type:"default"},{labelName:"Created",name:"created",type:"datetime"}]},{cloudServiceProvider:"k8s",labelName:"ConfigMap",entityName:"config_map",entityColumn:[{labelName:"Namespace",name:"namespace",type:"default"},{labelName:"Name",name:"name",type:"default"},{labelName:"Created",name:"created",type:"datetime"}]},{cloudServiceProvider:"k8s",labelName:"Secret",entityName:"secret",entityColumn:[{labelName:"Namespace",name:"namespace",type:"default"},{labelName:"Name",name:"name",type:"default"},{labelName:"Type",name:"secret_type",type:"default"},{labelName:"Created",name:"created",type:"datetime"}]},{cloudServiceProvider:"k8s",labelName:"Role",entityName:"role",entityColumn:[{labelName:"Namespace",name:"namespace",type:"default"},{labelName:"Name",name:"name",type:"default"},{labelName:"Created",name:"created",type:"datetime"}]},{cloudServiceProvider:"k8s",labelName:"Role Binding",entityName:"role_binding",entityColumn:[{labelName:"Namespace",name:"namespace",type:"default"},{labelName:"Name",name:"name",type:"default"},{labelName:"Role",name:"role_ref",type:"default"},{labelName:"Created",name:"created",type:"datetime"}]},{cloudServiceProvider:"k8s",labelName:"Cluster Role",entityName:"cluster_role",entityColumn:[{labelName:"Name",name:"name",type:"default"},{labelName:"Created",name:"created",type:"datetime"}]},{cloudServiceProvider:"k8s",labelName:"Cluster Role Binding",entityName:"cluster_role_binding",entityColumn:[{labelName:"Name",name:"name",type:"default"},{labelName:"Cluster Role",name:"role_ref",type:"default"},{labelName:"Created",name:"created",type:"datetime"}]},{cloudServiceProvider:"k8s",labelName:"Persistent Volume",entityName:"persistent_volume",entityColumn:[{labelName:"Name",name:"name",type:"default"},{labelName:"Capacity",name:"capacity",type:"default"},{labelName:"Access Modes",name:"access_modes",type:"default"},{labelName:"Reclaim Policy",name:"reclaim_policy",type:"default"},{labelName:"Status",name:"phase",type:"default"},{labelName:"Request",name:"claim_ref",type:"default"},{labelName:"Storage Class",name:"storage_class_name",type:"default"},{labelName:"Reason",name:"reason",type:"default"},{labelName:"Created",name:"created",type:"datetime"}]},{cloudServiceProvider:"k8s",labelName:"Persistent Volume Claim",entityName:"persistent_volume_claim",entityColumn:[{labelName:"Namespace",name:"namespace",type:"default"},{labelName:"Name",name:"name",type:"default"},{labelName:"Phase",name:"phase",type:"default"},{labelName:"VolumeName",name:"volume_name",type:"default"},{labelName:"Capacity",name:"capacity",type:"default"},{labelName:"Request",name:"request",type:"default"},{labelName:"Access Mode",name:"access_mode",type:"default"},{labelName:"Storage Class",name:"storage_class",type:"default"},{labelName:"Created",name:"created",type:"datetime"}]},{cloudServiceProvider:"k8s",labelName:"Storage Class",entityName:"storage_class",entityColumn:[{labelName:"Name",name:"name",type:"default"},{labelName:"Provisioner",name:"provisioner",type:"default"},{labelName:"Parameters",name:"parameters",type:"key-value"},{labelName:"Created",name:"created",type:"datetime"}]},{cloudServiceProvider:"k8s",labelName:"StatefulSet",entityName:"stateful_set",entityColumn:[{labelName:"Namespace",name:"namespace",type:"default"},{labelName:"Name",name:"name",type:"default"},{labelName:"Created",name:"created",type:"datetime"}]},{cloudServiceProvider:"k8s",labelName:"Ingress",entityName:"ingress",entityColumn:[{labelName:"Namespace",name:"namespace",type:"default"},{labelName:"Name",name:"name",type:"default"},{labelName:"Created",name:"created",type:"datetime"}]},{cloudServiceProvider:"k8s",labelName:"DaemonSet",entityName:"daemon_set",entityColumn:[{labelName:"Namespace",name:"namespace",type:"default"},{labelName:"Name",name:"name",type:"default"},{labelName:"Created",name:"created",type:"datetime"}]},{cloudServiceProvider:"k8s",labelName:"Endpoint",entityName:"endpoint",entityColumn:[{labelName:"Namespace",name:"namespace",type:"default"},{labelName:"Name",name:"name",type:"default"},{labelName:"Node",name:"node_name",type:"default"},{labelName:"Created",name:"created",type:"datetime"}]},{cloudServiceProvider:"k8s",labelName:"Event",entityName:"event",entityColumn:[{labelName:"Type",name:"k8s_event_type",type:"default"},{labelName:"Reason",name:"reason",type:"default"},{labelName:"Object Kind",name:"object_kind",type:"default"},{labelName:"Object Name",name:"object_name",type:"default"},{labelName:"Message",name:"message",type:"default"},{labelName:"Last Time Stamp",name:"time_stamp",type:"datetime"},{labelName:"Created",name:"created",type:"datetime"}]},{cloudServiceProvider:"k8s",labelName:"API Service",entityName:"api_service",entityColumn:[{labelName:"Name",name:"name",type:"default"},{labelName:"Created",name:"created",type:"datetime"}]},{cloudServiceProvider:"k8s",labelName:"ServiceAccount",entityName:"service_account",entityColumn:[{labelName:"Namespace",name:"namespace",type:"default"},{labelName:"Name",name:"name",type:"default"},{labelName:"Created",name:"created",type:"datetime"}]},{cloudServiceProvider:"k8s",labelName:"Horizontal Pod Autoscaler",entityName:"horizontal_pod_autoscaler",entityColumn:[{labelName:"Name",name:"name",type:"default"},{labelName:"Namespace",name:"namespace",type:"default"},{labelName:"Scale Target",name:"scale_target",type:"default"},{labelName:"Target CPU Utilization Percentage",name:"target_cpu_utilization_percentage",type:"default"}]},{cloudServiceProvider:"k8s",labelName:"Schedule",entityName:"schedule",entityColumn:[{labelName:"Name",name:"name",type:"default"},{labelName:"Kind",name:"kind",type:"default"},{labelName:"Namespace Name",name:"namespace_name",type:"default"},{labelName:"Resource Name",name:"resource_name",type:"default"},{labelName:"Launch Template Name",name:"launch_template_name",type:"default"},{labelName:"State",name:"state",type:"default"},{labelName:"Start",name:"start_time",type:"default"},{labelName:"Stop",name:"stop_time",type:"default"},{labelName:"Created",name:"created",type:"datetime"}]}],b=[{cloudServiceProvider:"aws_cloud",name:"ALL",labelName:"AWS Cloud (ALL)"},{cloudServiceProvider:"k8s",name:"ALL",labelName:"K8s (ALL)"}],f=[{labelName:"Name",name:"name",type:"default"},{labelName:"AMI Name",name:"field_image_id",type:"default"},{labelName:"Instance type",name:"field_instance_type",type:"default"},{labelName:"VPC",name:"field_vpc",type:"join",info:{entityTypeId:"aws_cloud_vpc",keyColumn:"vpc_id",valueColumn:"name"}},{labelName:"Max count",name:"field_max_count",type:"default"},{labelName:"Status",name:"field_workflow_status",type:"default"}],y=[{labelName:"Name",name:"name",type:"default"},{labelName:"Namespace",name:"field_namespace",type:"default"},{labelName:"Object",name:"field_object",type:"array"},{labelName:"Enable time scheduler",name:"field_enable_time_scheduler",type:"boolean",value:["On","Off"]},{labelName:"Workflow status",name:"field_workflow_status",type:"default"}],N=t(2),j=t.n(N),v=t(3),h=t(0),x=function(){var e=Object(v.a)(j.a.mark((function e(a,t,n){var r,l,c,o,i,s;return j.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return(r=new FormData).append("grant_type","authorization_code"),r.append("client_id",t),r.append("client_secret",u),r.append("code",a),r.append("redirect_uri",n),e.next=8,fetch("/oauth/token",{method:"POST",body:r});case 8:if((l=e.sent).ok){e.next=11;break}throw new Error("Authorization failed");case 11:return e.next=13,l.json();case 13:if(!("access_token"in(c=e.sent))){e.next=23;break}o=c.access_token,i=c.refresh_token,s=(new Date).getTime()+1e3*c.expires_in,window.localStorage.setItem("accessToken",o),window.localStorage.setItem("refreshToken",i),window.localStorage.setItem("expiresDatetime","".concat(s)),e.next=24;break;case 23:throw new Error("Authorization failed");case 24:case"end":return e.stop()}}),e)})));return function(a,t,n){return e.apply(this,arguments)}}(),g=function(){var e=Object(v.a)(j.a.mark((function e(){var a,t,n,r,l,c;return j.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return console.group("Authorization Code Grant"),e.next=3,fetch("/clouds/cloud_dashboard/config/client_id");case 3:return a=e.sent,e.next=6,fetch("/clouds/cloud_dashboard/config/callback_uri");case 6:if(t=e.sent,a.ok&&t.ok){e.next=13;break}return console.log("Client ID : No"),console.error("Authorization failed."),console.groupEnd(),window.location.href=m,e.abrupt("return");case 13:return e.next=15,a.json();case 15:return n=e.sent.id,e.next=18,t.json();case 18:if(r=e.sent.uri,console.log("Client ID : Yes"),l=new URL(window.location.href).searchParams,null!==(c=l.get("code"))){e.next=28;break}return console.log("Authorization code : No"),console.error("Authorization failed."),console.groupEnd(),window.location.href=m,e.abrupt("return");case 28:return console.log("authorization code : Yes"),e.prev=29,e.next=32,x(c,n,r);case 32:console.log("Access token : Yes"),console.groupEnd(),window.location.href=m+"/aws_cloud/instance",e.next=43;break;case 37:e.prev=37,e.t0=e.catch(29),console.log("Access token : No"),console.error("Authorization failed."),console.groupEnd(),window.location.href=m;case 43:case"end":return e.stop()}}),e,null,[[29,37]])})));return function(){return e.apply(this,arguments)}}(),O=function(){return Object(n.useEffect)((function(){g()}),[]),Object(h.jsx)("div",{className:"container-fluid",children:Object(h.jsx)("div",{className:"row",children:Object(h.jsx)("div",{className:"col",children:Object(h.jsx)("h2",{children:"Please waiting..."})})})})},_=t(4),C=function(){var e=Object(n.useState)(""),a=Object(_.a)(e,2),t=a[0],r=a[1],l=Object(n.useState)(""),c=Object(_.a)(l,2),o=c[0],i=c[1],s=function(){var e=Object(v.a)(j.a.mark((function e(){var a,t;return j.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return window.localStorage.removeItem("jsonapiServerUri"),e.next=3,fetch("/clouds/cloud_dashboard/config/client_id");case 3:return a=e.sent,e.next=6,fetch("/clouds/cloud_dashboard/config/callback_uri");case 6:if(t=e.sent,!a.ok||!t.ok){e.next=18;break}return e.t0=r,e.next=11,a.json();case 11:return e.t1=e.sent.id,(0,e.t0)(e.t1),e.t2=i,e.next=16,t.json();case 16:e.t3=e.sent.uri,(0,e.t2)(e.t3);case 18:case"end":return e.stop()}}),e)})));return function(){return e.apply(this,arguments)}}();return Object(n.useEffect)((function(){s()}),[]),Object(h.jsx)("div",{className:"container-fluid",children:Object(h.jsx)("div",{className:"row",children:Object(h.jsx)("div",{className:"col",children:Object(h.jsx)("form",{children:Object(h.jsx)("button",{type:"button",className:"btn btn-default",onClick:function(){var e="/oauth/authorize?response_type=code&client_id=".concat(t,"&redirect_uri=").concat(o);window.location.href=e},disabled:""===t,children:"login"})})})})})},k=t(31),w=t(32),S=function(){function e(){Object(k.a)(this,e),this.cache={}}return Object(w.a)(e,[{key:"getJson",value:function(){var e=Object(v.a)(j.a.mark((function e(a,t){var n,r,l,c;return j.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:if(!(a in this.cache)){e.next=5;break}if(n=this.cache[a],!((r=(new Date).getTime())>=n.unixtime&&r-n.unixtime<864e5)){e.next=5;break}return e.abrupt("return",n.response);case 5:return e.next=7,fetch(a,t);case 7:return e.next=9,e.sent.json();case 9:return l=e.sent,c=(new Date).getTime(),this.cache[a]={response:l,unixtime:c},e.abrupt("return",l);case 13:case"end":return e.stop()}}),e,this)})));return function(a,t){return e.apply(this,arguments)}}()}],[{key:"getInstance",value:function(){return this.instance}}]),e}();S.instance=new S;var P=S,I=t(61),L=function(){var e=Object(n.useState)(function(){var e=localStorage.getItem("cloudContext");return null!==e?JSON.parse(e):b[0]}()),a=Object(_.a)(e,2),t=a[0],r=a[1],l=Object(n.useState)(Object(o.a)(b)),c=Object(_.a)(l,2),i=c[0],s=c[1],u=Object(I.a)().i18n;Object(n.useEffect)((function(){var e=window.localStorage.getItem("languageName");null!==e&&u.changeLanguage(e)}),[]),Object(n.useEffect)((function(){(function(){var e=Object(v.a)(j.a.mark((function e(){var a,n,l,c,i;return j.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:a=["aws_cloud","k8s"],n=Object(o.a)(b),l=j.a.mark((function e(){var a,t;return j.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return a=i[c],e.next=3,P.getInstance().getJson("/jsonapi/cloud_config/".concat(a));case 3:t=e.sent.data.map((function(e){return{cloudServiceProvider:a,name:e.attributes.cloud_context,labelName:e.attributes.name}})),n=[].concat(Object(o.a)(n),Object(o.a)(t));case 5:case"end":return e.stop()}}),e)})),c=0,i=a;case 4:if(!(c<i.length)){e.next=9;break}return e.delegateYield(l(),"t0",6);case 6:c++,e.next=4;break;case 9:s(n),0===n.filter((function(e){return e.cloudServiceProvider===t.cloudServiceProvider&&e.name===t.name})).length&&r(b[0]);case 11:case"end":return e.stop()}}),e)})));return function(){return e.apply(this,arguments)}})()()}),[]);return{cloudContext:t,cloudContextList:i,dispatch:function(e){switch(e.type){case"setCloudContext":localStorage.setItem("cloudContext",JSON.stringify(e.message)),r(e.message);break;case"setLanguage":u.changeLanguage(e.message),window.localStorage.setItem("languageName",e.message)}}}},A=Object(n.createContext)({cloudContext:b[0],cloudContextList:[],dispatch:function(){}}),T=t(13),D=t(10),R=t.n(D),E=function(e,a){return"000000".concat(e).slice(-a)},U=function(e,a){for(var t=1,n=0;n<a;n+=1)t*=10;if(Math.floor(e)!==e)return"".concat(Math.round(e*t)/t);var r="".concat(e);if(0!==a){r+=".";for(var l=0;l<a;l+=1)r+="0"}return r},z=function(e,a,t){if(null===e)return"";switch(a.type){case"datetime":return function(e){var a=new Date(e),t=a.getFullYear(),n=a.getMonth()+1,r=a.getDate(),l=a.getHours(),c=a.getMinutes(),o="".concat(t,"/").concat(E(n,2),"/").concat(E(r,2));return o+" - ".concat(E(l,2),":").concat(E(c,2))}(e);case"memory":return e>=1073741824?"".concat(U(e/1073741824,2),"Gi"):e>=1048576?"".concat(U(e/1048576,2),"Mi"):e>=1024?"".concat(U(e/1024,2),"Ki"):U(e,2);case"key-value":var n,r=[],l=Object(T.a)(e);try{for(l.s();!(n=l.n()).done;){var c=n.value;r.push("".concat(c.item_key,":").concat(c.item_value))}}catch(u){l.e(u)}finally{l.f()}return r.join(", ");case"cost":return"$".concat(e);case"boolean":return e?a.value[0]:a.value[1];case"array":return e.map((function(e){return"".concat(e)})).join(", ");case"join":var o,i=Object(T.a)(t[a.info.entityTypeId]);try{for(i.s();!(o=i.n()).done;){var s=o.value;if(s.attributes[a.info.keyColumn]===e)return s.attributes[a.info.valueColumn]}}catch(u){i.e(u)}finally{i.f()}return"(".concat(e,")");default:return e}},V=function(e){return"/".concat(e.cloudServiceProvider,"/").concat(e.entityName)},M=function(e){return"".concat(e.cloudServiceProvider,"_").concat(e.entityName)},F=function(){var e=Object(v.a)(j.a.mark((function e(a,t){var n,r,l,c,o;return j.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:for(r in(n=[]).push({key:"page[limit]",value:"".concat(t.limit)}),n.push({key:"page[offset]",value:"".concat(t.offset)}),t.filter)n.push({key:"filter[".concat(r,"]"),value:t.filter[r]});return""!==t.sort.key&&n.push("ASC"===t.sort.direction?{key:"sort",value:t.sort.key}:{key:"sort",value:"-"+t.sort.key}),l="/jsonapi/".concat(a,"/").concat(a),n.length>0&&(l+="?"+n.map((function(e){return e.key+"="+e.value})).join("&")),null!==(c=window.localStorage.getItem("jsonapiServerUri"))&&(l=c+l),e.next=11,P.getInstance().getJson(l);case 11:return o=e.sent,e.abrupt("return",o.data);case 13:case"end":return e.stop()}}),e)})));return function(a,t){return e.apply(this,arguments)}}(),q=function(){var e=Object(v.a)(j.a.mark((function e(a){var t,n,r,l,c,i,s,u,m=arguments;return j.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:for(r in t=m.length>1&&void 0!==m[1]?m[1]:{},n=[],t)n.push({key:"filter[".concat(r,"]"),value:t[r]});l="/jsonapi/".concat(a,"/").concat(a),n.length>0&&(l+="?"+n.map((function(e){return e.key+"="+e.value})).join("&")),null!==(c=window.localStorage.getItem("jsonapiServerUri"))&&(l=c+l),i=[],s=P.getInstance();case 9:return e.next=12,s.getJson(l);case 12:if(u=e.sent,i=[].concat(Object(o.a)(i),Object(o.a)(u.data)),void 0===u.links.next){e.next=18;break}l=u.links.next.href,e.next=19;break;case 18:return e.abrupt("break",21);case 19:e.next=9;break;case 21:return e.abrupt("return",i);case 22:case"end":return e.stop()}}),e)})));return function(a){return e.apply(this,arguments)}}(),G=function(e){return"ALL"===e.name?"/".concat(e.cloudServiceProvider,"/server_template"):"/server_template/".concat(e.cloudServiceProvider,"/").concat(e.name)},J=function(){var e=Object(v.a)(j.a.mark((function e(a){var t,n,r,l,c;return j.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:t={},n=Object(T.a)(a),e.prev=2,n.s();case 4:if((r=n.n()).done){e.next=16;break}if("join"===(l=r.value).type){e.next=8;break}return e.abrupt("continue",14);case 8:if(!((c=l.info.entityTypeId)in t)){e.next=11;break}return e.abrupt("continue",14);case 11:return e.next=13,q(c);case 13:t[c]=e.sent;case 14:e.next=4;break;case 16:e.next=21;break;case 18:e.prev=18,e.t0=e.catch(2),n.e(e.t0);case 21:return e.prev=21,n.f(),e.finish(21);case 24:return e.abrupt("return",t);case 25:case"end":return e.stop()}}),e,null,[[2,18,21,24]])})));return function(a){return e.apply(this,arguments)}}(),K=function(e,a,t,n){var r,l=[],c=Object(T.a)(e);try{for(c.s();!(r=c.n()).done;){var o,i=r.value,s={id:i.id,value:{}},u=Object(T.a)(a);try{for(u.s();!(o=u.n()).done;){var m=o.value,d=i.attributes[m.name];s.value[m.name]="cloud_context"===m.name?t.labelName:z(d,m,n)}}catch(p){u.e(p)}finally{u.f()}l.push(s)}}catch(p){c.e(p)}finally{c.f()}return l},H=function(){var e=Object(v.a)(j.a.mark((function e(){var a,t,n;return j.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return a="/clouds/cloud_config_location",null!==(t=window.localStorage.getItem("jsonapiServerUri"))&&(a=t+a),n=P.getInstance(),e.next=6,n.getJson(a);case 6:return e.abrupt("return",e.sent);case 7:case"end":return e.stop()}}),e)})));return function(){return e.apply(this,arguments)}}(),Y=function(e,a){var t,n={},r=Object(T.a)(e);try{var l=function(){var e=t.value,r="".concat(e.Latitude,",").concat(e.Longitude);r in n?n[r].icon=new R.a.Icon({iconUrl:a}):n[r]={icon:new R.a.Icon({iconUrl:e.Items[0].Image}),cloudServiceProvider:e.Type,position:[parseFloat(e.Latitude),parseFloat(e.Longitude)],item:[]};var l=e.Items.map((function(a){return{iconUrl:a.Image,entityViewUrl:"aws_cloud"===e.Type?"/aws_cloud/instance":"/k8s/node",name:a.Name,positionLabel:"".concat(e.City,", ").concat(e.Country)}}));n[r].item=[].concat(Object(o.a)(n[r].item),Object(o.a)(l))};for(r.s();!(t=r.n()).done;)l()}catch(c){r.e(c)}finally{r.f()}return Object.values(n)},B=function(e){var a=e.value,t=e.setvalue,n=e.dataList,r=Object(I.a)().t;return Object(h.jsxs)("select",{className:"form-select form-control",value:a,onChange:function(e){t(e.currentTarget.value)},children:[Object(h.jsxs)("option",{value:"",children:["- ",r("All")," -"]}),n.map((function(e){return Object(h.jsx)("option",{value:e,children:e},e)}))]})},Z=function(){var e=Object(v.a)(j.a.mark((function e(a,t){var n,r,l;return j.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return n={},"ALL"!==a.name&&(n.cloud_context=a.name),r="".concat(a.cloudServiceProvider,"_").concat(t),e.next=5,q(r,n);case 5:return l=e.sent,e.abrupt("return",l.map((function(e){return e.attributes.name})));case 7:case"end":return e.stop()}}),e)})));return function(a,t){return e.apply(this,arguments)}}(),W=function(e){var a=e.columnKey,t=e.columnName,r=e.setColumnName,l=Object(n.useContext)(A).cloudContext,c=Object(n.useState)([]),o=Object(_.a)(c,2),i=o[0],s=o[1];return Object(n.useEffect)((function(){Z(l,a).then((function(e){s(e)}))}),[l]),Object(h.jsx)(B,{value:t,dataList:i,setvalue:function(e){r(e)}})},Q=function(e){var a=e.children;return Object(h.jsx)("label",{className:"control-label",children:a})},$=function(){var e=Object(v.a)(j.a.mark((function e(a,t,n,r){var l,c;return j.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return l="ALL"===a.name?"/cloud_dashboard/".concat(a.cloudServiceProvider,"/").concat(t,"/entity/count"):"/cloud_dashboard/".concat(a.cloudServiceProvider,"/").concat(a.name,"/").concat(t,"/entity/count"),""!==n&&(l+="?namespace=".concat(n)),""!==r&&(l+="?namespace_name=".concat(r)),null!==(c=window.localStorage.getItem("jsonapiServerUri"))&&(l=c+l),e.next=7,P.getInstance().getJson(l);case 7:return e.abrupt("return",e.sent.count);case 8:case"end":return e.stop()}}),e)})));return function(a,t,n,r){return e.apply(this,arguments)}}(),X=function(e){var a=e.entityTypeId,t=e.namespace,r=e.namespaceName,l=e.itemCount,c=e.setItemCount,o=Object(n.useContext)(A).cloudContext,i=Object(I.a)().t;return Object(n.useEffect)((function(){$(o,a,t,r).then((function(e){c(e)}))}),[o,t,r]),Object(h.jsxs)(Q,{children:[i("ItemCount"),": ",l]})},ee=function(e){var a=e.ariaLabel,t=e.onClick,n=e.children;return Object(h.jsx)("span",{"aria-label":a,onClick:t,children:Object(h.jsx)("span",{"aria-hidden":"true",children:n})})},ae=function(e){var a=e.pageIndex,t=e.setPageIndex,n=e.itemCount,r=Math.floor(1*(n+30-1)/30),l=0===a,c=0===n||a===r-1;return Object(h.jsx)("nav",{children:Object(h.jsxs)("ul",{className:"pagination",children:[Object(h.jsx)("li",{className:l?"disabled":"",children:Object(h.jsx)(ee,{ariaLabel:"First",onClick:function(){t(0)},children:"\xab"})}),Object(h.jsx)("li",{className:l?"disabled":"",children:Object(h.jsx)(ee,{ariaLabel:"Previous",onClick:function(){t(Math.max(0,a-1))},children:"\uff1c"})}),Object(h.jsx)("li",{children:Object(h.jsx)("span",{children:a+1})}),Object(h.jsx)("li",{className:c?"disabled":"",children:Object(h.jsx)(ee,{ariaLabel:"Next",onClick:function(){t(Math.min(a+1,r-1))},children:"\uff1e"})}),Object(h.jsx)("li",{className:c?"disabled":"",children:Object(h.jsx)(ee,{ariaLabel:"Last",onClick:function(){t(r-1)},children:"\xbb"})})]})})},te=function(e){var a=e.direction,t="ASC"===a?"glyphicon-chevron-up":"glyphicon-chevron-down",n="ASC"===a?"descending":"ascending";return Object(h.jsx)("span",{className:"icon glyphicon ".concat(t," icon-after"),"aria-hidden":"true","data-toggle":"tooltip","data-placement":"bottom",title:"","data-original-title":"Sort by ".concat(n," order")})},ne=function(e){var a=e.dataColumn,t=e.sortInfo,n=e.setSortInfo,r=function(e){t.key===e?"ASC"!==t.direction?n({key:"",direction:"ASC"}):n({key:e,direction:"DESC"}):n({key:e,direction:"ASC"})};return t.key!==a.key?Object(h.jsx)("th",{onClick:function(){r(a.key)},children:a.label},a.key):Object(h.jsxs)("th",{className:"is-active",onClick:function(){r(a.key)},children:[a.label,Object(h.jsx)("a",{href:"#",onClick:function(e){e.preventDefault()},children:Object(h.jsx)(te,{direction:t.direction})})]},a.key)},re=function(e){var a=e.dataRecord,t=e.dataColumnList,n=e.className;return Object(h.jsx)("tr",{className:n,children:t.map((function(e){return Object(h.jsx)("td",{children:e.key in a.value?a.value[e.key]:""},e.key)}))},a.id)},le=function(e){var a=e.dataColumnList,t=e.dataRecordList,n=e.sortInfo,r=e.setSortInfo;return Object(h.jsx)("div",{className:"table-responsive",children:Object(h.jsxs)("table",{className:"responsive-enabled table table-hover table-striped",children:[Object(h.jsx)("thead",{children:Object(h.jsx)("tr",{children:a.map((function(e){return Object(h.jsx)(ne,{dataColumn:e,sortInfo:n,setSortInfo:r})}))})}),Object(h.jsx)("tbody",{children:t.map((function(e,t){return Object(h.jsx)(re,{dataRecord:e,dataColumnList:a,className:t%2===0?"odd":"even"})}))})]})})},ce=function(e){var a=e.entityTypeId,t=e.entityColumnList,r=e.namespace,l=e.namespaceName,c=e.pageIndex,o=Object(n.useContext)(A).cloudContext,i=Object(n.useState)([]),s=Object(_.a)(i,2),u=s[0],m=s[1],d=Object(n.useState)([]),p=Object(_.a)(d,2),b=p[0],f=p[1],y=Object(n.useState)({key:"",direction:"ASC"}),N=Object(_.a)(y,2),x=N[0],g=N[1];return Object(n.useEffect)((function(){(function(){var e=Object(v.a)(j.a.mark((function e(){var n,i,s,u,d;return j.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return n=t.map((function(e){return{key:e.name,label:e.labelName}})),m(n),e.next=4,J(t);case 4:return i=e.sent,s={},""!==r&&(s.namespace=r),""!==l&&(s.namespaceName=l),"ALL"!==o.name&&(s.cloud_context=o.name),u={limit:30,offset:30*c,filter:s,sort:x},e.next=12,F(a,u);case 12:d=e.sent,f(K(d,t,o,i));case 14:case"end":return e.stop()}}),e)})));return function(){return e.apply(this,arguments)}})()()}),[o,a,t,x,r,l,c]),Object(h.jsx)(le,{dataColumnList:u,dataRecordList:b,sortInfo:x,setSortInfo:g})},oe=function(e){var a=e.entityTypeId,t=e.column,r=Object(n.useState)(""),l=Object(_.a)(r,2),c=l[0],o=l[1],i=Object(n.useState)(""),s=Object(_.a)(i,2),u=s[0],m=s[1],d=Object(n.useState)(0),p=Object(_.a)(d,2),b=p[0],f=p[1],y=Object(n.useState)(0),N=Object(_.a)(y,2),j=N[0],v=N[1];Object(n.useEffect)((function(){""!==c&&m("")}),[c]),Object(n.useEffect)((function(){""!==u&&o("")}),[u]);var x=t.map((function(e){return e.labelName})).includes("Namespace"),g=t.map((function(e){return e.labelName})).includes("Namespace Name");return Object(h.jsxs)(h.Fragment,{children:[x?Object(h.jsxs)("div",{style:{marginTop:"2rem"},children:[Object(h.jsx)("label",{className:"control-label",children:"Namespace"}),Object(h.jsx)(W,{columnKey:"namespace",columnName:c,setColumnName:o})]}):Object(h.jsx)(h.Fragment,{}),g?Object(h.jsxs)("div",{style:{marginTop:"2rem"},children:[Object(h.jsx)("label",{className:"control-label",children:"Namespace Name"}),Object(h.jsx)(W,{columnKey:"namespace_name",columnName:u,setColumnName:m})]}):Object(h.jsx)(h.Fragment,{}),Object(h.jsx)("div",{style:{marginTop:"2rem"},children:Object(h.jsx)(X,{entityTypeId:a,namespace:c,namespaceName:u,itemCount:b,setItemCount:f})}),Object(h.jsx)("div",{style:{marginTop:"2rem"},children:Object(h.jsx)(ce,{entityTypeId:a,entityColumnList:t,namespace:c,namespaceName:u,pageIndex:j})}),Object(h.jsx)(ae,{pageIndex:j,setPageIndex:v,itemCount:b})]})},ie=function(e){var a=e.to,t=e.onClick,n=e.className,r=e.children;return Object(h.jsx)("li",{className:n,children:Object(h.jsx)(i.b,{to:a,onClick:t,children:r})})},se=function(e){var a=e.href,t=e.className,n=e.children;return Object(h.jsx)("li",{className:t,children:Object(h.jsx)("a",{href:a,children:n})})},ue=function(e){var a=e.menuName,t=e.linkPropsList;return Object(h.jsxs)("li",{className:"dropdown",children:[Object(h.jsxs)("a",{href:"#",className:"dropdown-toggle","data-toggle":"dropdown",children:["".concat(a)," ",Object(h.jsx)("span",{className:"caret"})]}),Object(h.jsx)("ul",{className:"dropdown-menu",role:"menu",children:t.map((function(e,a){return Object(h.jsx)(ie,{to:e.to,onClick:e.onClick,children:e.children},a)}))})]})},me=function(e){var a=e.anchorPropsList;return Object(h.jsx)("ul",{className:"dropdown-menu",role:"menu",children:a.map((function(e,a){return Object(h.jsx)(se,{href:e.href,children:e.children},a)}))})},de=function(){var e=Object(v.a)(j.a.mark((function e(a,t){var n,r,l,c,o,i;return j.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return(n=new FormData).append("grant_type","refresh_token"),n.append("client_id",a),n.append("client_secret",u),n.append("refresh_token",t),e.next=7,fetch("/oauth/token",{method:"POST",body:n});case 7:if((r=e.sent).ok){e.next=10;break}throw new Error("Refresh failed");case 10:return e.next=12,r.json();case 12:if(!("access_token"in(l=e.sent))){e.next=22;break}c=l.access_token,o=l.refresh_token,i=(new Date).getTime()+1e3*l.expires_in,window.localStorage.setItem("accessToken",c),window.localStorage.setItem("refreshToken",o),window.localStorage.setItem("expiresDatetime","".concat(i)),e.next=23;break;case 22:throw new Error("Refresh failed");case 23:case"end":return e.stop()}}),e)})));return function(a,t){return e.apply(this,arguments)}}(),pe=function(){var e=Object(v.a)(j.a.mark((function e(){var a,t,n,r,l;return j.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:if(a=window.localStorage.getItem("accessToken"),t=window.localStorage.getItem("expiresDatetime"),null!==a&&null!=t){e.next=5;break}return window.location.href=m,e.abrupt("return");case 5:return console.group("Authorization Code Grant"),e.next=8,fetch("/clouds/cloud_dashboard/config/client_id");case 8:if(n=e.sent,r=window.localStorage.getItem("refreshToken"),n.ok&&null!==r){e.next=16;break}return console.log("Client ID : No"),console.error("Authorization failed."),console.groupEnd(),window.location.href=m,e.abrupt("return");case 16:return e.next=18,n.json();case 18:if(l=e.sent.id,console.log("Client ID : Yes"),!((new Date).getTime()<=parseInt(t,10))){e.next=25;break}return console.log("Token expired : No"),console.groupEnd(),e.abrupt("return");case 25:console.log("Token expired : Yes"),de(l,r).then((function(){console.log("Access token : Yes"),console.groupEnd()})).catch((function(){console.log("Access token : No"),console.error("Authorization failed."),console.groupEnd(),window.location.href=m}));case 27:case"end":return e.stop()}}),e)})));return function(){return e.apply(this,arguments)}}(),be=function(){var e=Object(n.useContext)(A),a=e.cloudContext,t=e.cloudContextList,r=e.dispatch,l=Object(I.a)().t;Object(n.useEffect)((function(){(function(){var e=Object(v.a)(j.a.mark((function e(){var a;return j.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return e.next=2,pe();case 2:if(null!==window.localStorage.getItem("jsonapiServerUri")){e.next=12;break}return e.next=5,fetch("/clouds/cloud_dashboard/config/jsonapi_server_uri");case 5:if(!(a=e.sent).ok){e.next=12;break}return e.t0=window.localStorage,e.next=10,a.json();case 10:e.t1=e.sent.uri,e.t0.setItem.call(e.t0,"jsonapiServerUri",e.t1);case 12:case"end":return e.stop()}}),e)})));return function(){return e.apply(this,arguments)}})()()}),[]);var c=function(e){r({type:"setCloudContext",message:e})};return Object(h.jsx)("header",{className:"navbar navbar-default navbar-fixed-top",role:"banner",children:Object(h.jsxs)("div",{className:"container-fluid",children:[Object(h.jsxs)("div",{className:"navbar-header",children:[Object(h.jsx)("div",{className:"region region-navigation",children:Object(h.jsx)("a",{className:"logo navbar-btn pull-left",href:"/",title:l("Home"),rel:"home",children:Object(h.jsx)("img",{src:"/themes/contrib/bootstrap_cloud/logo.svg",alt:l("Home")})})}),Object(h.jsxs)("button",{type:"button",className:"navbar-toggle","data-toggle":"collapse","data-target":"#navbar-collapse",children:[Object(h.jsx)("span",{className:"sr-only",children:"Toggle navigation"}),Object(h.jsx)("span",{className:"icon-bar"}),Object(h.jsx)("span",{className:"icon-bar"}),Object(h.jsx)("span",{className:"icon-bar"})]})]}),Object(h.jsx)("div",{className:"navbar-collapse collapse",children:Object(h.jsxs)("div",{className:"region region-navigation-collapsible",children:[Object(h.jsxs)("nav",{role:"navigation",className:"contextual-region",children:[Object(h.jsx)("h2",{className:"sr-only",children:"Main navigation"}),Object(h.jsxs)("ul",{className:"nav navbar-nav",role:"menu",children:[Object(h.jsx)(ie,{to:"/providers",children:l("Home")}),Object(h.jsx)(ue,{menuName:a.labelName,linkPropsList:t.map((function(e){var a="aws_cloud"===e.cloudServiceProvider?d:p;return{to:V(a[0]),onClick:function(){return c(e)},children:e.labelName}}))}),Object(h.jsx)(ue,{menuName:"Launch template",linkPropsList:t.map((function(e){return{to:G(e),onClick:function(){return c(e)},children:e.labelName}}))}),Object(h.jsxs)("li",{className:"dropdown",children:[Object(h.jsxs)("a",{href:"/clouds/design",className:"dropdown-toggle","data-toggle":"dropdown",children:["Design ",Object(h.jsx)("span",{className:"caret"})]}),Object(h.jsxs)("ul",{className:"dropdown-menu",role:"menu",children:[Object(h.jsxs)("li",{className:"dropdown-submenu",children:[Object(h.jsxs)("a",{href:"/clouds",className:"dropdown-submenu-toggle",children:["Projects ",Object(h.jsx)("span",{className:"caret"})]}),Object(h.jsx)(me,{anchorPropsList:t.filter((function(e){return"k8s"===e.cloudServiceProvider&&"ALL"!==e.name})).map((function(e){return{href:"/clouds/design/project/".concat(e.name),children:e.labelName}}))})]}),Object(h.jsxs)("li",{className:"dropdown-submenu",children:[Object(h.jsxs)("a",{href:"/clouds",className:"dropdown-submenu-toggle",children:["Stores ",Object(h.jsx)("span",{className:"caret"})]}),Object(h.jsx)(me,{anchorPropsList:[{href:"/clouds/design/store/k8s_cost_store",children:"K8s cost store"},{href:"/clouds/design/store/k8s_namespace_resource_store",children:"K8s namespace resource store"},{href:"/clouds/design/store/k8s_node_resource_store",children:"K8s node resource store"},{href:"/clouds/design/store/k8s_pod_resource_store",children:"K8s pod resource store"}]})]})]})]}),Object(h.jsx)(se,{href:"/admin/structure/cloud_config",children:l("CloudConfig")})]})]}),Object(h.jsxs)("nav",{role:"navigation",className:"contextual-region",children:[Object(h.jsx)("h2",{className:"sr-only",children:"User account menu"}),Object(h.jsxs)("ul",{className:"menu menu--account nav navbar-nav navbar-right",children:[Object(h.jsx)(se,{className:"first",href:"/user",children:l("User")}),Object(h.jsx)(ie,{className:"last",to:"/",onClick:function(){window.localStorage.removeItem("accessToken"),window.localStorage.removeItem("refreshToken"),window.localStorage.removeItem("expiresDatetime")},children:l("Logout")})]})]})]})})]})})},fe=t(26),ye=t(19),Ne=[{key:"en",value:"English"},{key:"ja",value:"\u65e5\u672c\u8a9e"}];fe.a.use(ye.e).init({resources:{en:{translation:{Language:"Language",CloudConfig:"Cloud Config",languageName:"English",Home:"Home",User:"User",Logout:"Logout",ItemCount:"Item Count",All:"All"}},ja:{translation:{Language:"\u4f7f\u7528\u8a00\u8a9e",CloudConfig:"\u7ba1\u7406",languageName:"\u65e5\u672c\u8a9e",Home:"\u30db\u30fc\u30e0",User:"\u30e6\u30fc\u30b6\u30fc\u60c5\u5831",Logout:"\u30ed\u30b0\u30a2\u30a6\u30c8",ItemCount:"\u4ef6\u6570",All:"\u3059\u3079\u3066"}}},lng:"en",interpolation:{escapeValue:!1}});fe.a;var je=t(60),ve=t(57),he=function(e){return e.map((function(e){return[e[1],e[0]]}))},xe={fillColor:"#ddc",fillOpacity:1,color:"#ddc",opacity:1,weight:1},ge=function(e){var a=e.coordinates;return Object(h.jsx)(h.Fragment,{children:a.map((function(e,a){return Object(h.jsx)(ve.a,{pathOptions:xe,positions:he(e)},a)}))})},Oe=function(e){var a=e.feature;return"Polygon"===a.geometry.type?Object(h.jsx)(ge,{coordinates:a.geometry.coordinates}):Object(h.jsx)(h.Fragment,{children:a.geometry.coordinates.map((function(e,a){return Object(h.jsx)(ge,{coordinates:e},a)}))})},_e=function(e){var a=e.mapData;return Object(h.jsx)(h.Fragment,{children:a.features.map((function(e,a){return Object(h.jsx)(Oe,{feature:e},a)}))})},Ce=t(58),ke=t(59),we=function(e){var a=e.cloudServiceProvider,t=e.label,r=Object(n.useContext)(A),l=r.cloudContextList,c=r.dispatch,o=function(e){var t=l.filter((function(t){return t.cloudServiceProvider===a&&t.labelName===e}));t.length>=1&&c({type:"setCloudContext",message:t[0]})};return Object(h.jsxs)("div",{children:[Object(h.jsx)("div",{children:Object(h.jsx)(i.b,{to:t.entityViewUrl,target:"_blank",rel:"noreferrer",onClick:function(){o(t.name)},children:Object(h.jsx)("img",{src:t.iconUrl,alt:t.name})})}),Object(h.jsxs)("div",{children:[Object(h.jsx)("strong",{children:Object(h.jsx)(i.b,{to:t.entityViewUrl,target:"_blank",rel:"noreferrer",onClick:function(){o(t.name)},children:t.name})}),Object(h.jsx)("br",{}),Object(h.jsxs)("span",{className:"location",children:[Object(h.jsx)("span",{className:"glyphicon glyphicon-map-marker"}),t.positionLabel]})]})]})},Se=function(e){var a=e.item;return Object(h.jsx)(Ce.a,{position:a.position,icon:a.icon,children:Object(h.jsx)(ke.a,{children:a.item.map((function(e,t){return Object(h.jsx)(we,{cloudServiceProvider:a.cloudServiceProvider,label:e},t)}))})})},Pe=function(){var e=Object(v.a)(j.a.mark((function e(){var a,t,n,r,l;return j.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return e.next=2,fetch("/clouds/cloud_dashboard/config/marker_icon_uri");case 2:if((a=e.sent).ok){e.next=5;break}return e.abrupt("return",[]);case 5:return e.next=7,a.json();case 7:return t=e.sent,n=t.uri,e.next=11,H();case 11:return r=e.sent,l=Y(r,n),e.abrupt("return",l);case 14:case"end":return e.stop()}}),e)})));return function(){return e.apply(this,arguments)}}(),Ie=function(){var e=Object(v.a)(j.a.mark((function e(){var a,t;return j.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return e.next=2,fetch("/clouds/cloud_dashboard/config/map_geojson_uri");case 2:if((a=e.sent).ok){e.next=5;break}return e.abrupt("return",{features:[]});case 5:return e.next=7,a.json();case 7:return t=e.sent,e.next=10,P.getInstance().getJson(t.uri);case 10:return e.abrupt("return",e.sent);case 11:case"end":return e.stop()}}),e)})));return function(){return e.apply(this,arguments)}}(),Le=function(){var e=Object(n.useState)([]),a=Object(_.a)(e,2),t=a[0],r=a[1],l=Object(n.useState)({features:[]}),c=Object(_.a)(l,2),o=c[0],i=c[1];return Object(n.useEffect)((function(){(function(){var e=Object(v.a)(j.a.mark((function e(){return j.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return e.t0=r,e.next=3,Pe();case 3:return e.t1=e.sent,(0,e.t0)(e.t1),e.t2=i,e.next=8,Ie();case 8:e.t3=e.sent,(0,e.t2)(e.t3);case 10:case"end":return e.stop()}}),e)})));return function(){return e.apply(this,arguments)}})()()}),[]),Object(h.jsxs)(je.a,{center:[0,0],zoom:2,scrollWheelZoom:!1,style:{height:500,backgroundColor:"#4e5d6c"},children:[Object(h.jsx)(_e,{mapData:o}),t.map((function(e,a){return Object(h.jsx)(Se,{item:e},a)}))]})},Ae=function(){var e=Object(n.useContext)(A),a=e.cloudContextList,t=e.dispatch,r=Object(I.a)(),l=r.t,c=r.i18n;return Object(h.jsxs)("div",{className:"container-fluid",children:[Object(h.jsx)("div",{className:"row",children:Object(h.jsx)("div",{className:"col",children:Object(h.jsx)(Le,{})})}),Object(h.jsx)("div",{className:"row",children:Object(h.jsx)("div",{className:"col",children:Object(h.jsxs)("form",{children:[Object(h.jsxs)("div",{className:"form-group",style:{marginTop:"2rem"},children:[Object(h.jsx)("label",{className:"control-label",children:l("Language")}),Object(h.jsx)("br",{}),Object(h.jsx)("div",{className:"btn-group",role:"group",children:Ne.map((function(e){return Object(h.jsx)("button",{type:"button",className:"btn btn-default",disabled:c.language===e.key,onClick:function(){t({type:"setLanguage",message:e.key})},children:e.value})}))})]}),Object(h.jsx)("div",{className:"form-group",style:{marginTop:"2rem"},children:Object(h.jsx)("label",{className:"control-label",children:"Cloud Service Providers"})}),Object(h.jsx)("div",{className:"table-responsive",style:{marginTop:"2rem"},children:Object(h.jsxs)("table",{className:"table table-hover table-striped sticky-enabled sticky-table",children:[Object(h.jsx)("thead",{children:Object(h.jsx)("th",{children:"Name"})}),Object(h.jsx)("tbody",{children:a.filter((function(e){return"ALL"!==e.name})).map((function(e,a){return Object(h.jsx)("tr",{children:Object(h.jsx)("td",{children:e.labelName})},a)}))})]})})]})})})]})},Te=function(e){switch(e.cloudServiceProvider){case"aws_cloud":return"ALL"!==e.name?f:[{labelName:"Cloud Service Provider ID",name:"cloud_context",type:"default"}].concat(Object(o.a)(f));case"k8s":return"ALL"!==e.name?y:[{labelName:"Cloud Service Provider ID",name:"cloud_context",type:"default"}].concat(Object(o.a)(y))}},De=function(){var e=Object(v.a)(j.a.mark((function e(a,t){var n,r,l,c,o;return j.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return n=[],"ALL"!==a.name&&n.push({key:"filter[cloud_context]",value:a.name}),""!==t.key&&n.push("ASC"===t.direction?{key:"sort",value:t.key}:{key:"sort",value:"-"+t.key}),r="/jsonapi/cloud_launch_template/".concat(a.cloudServiceProvider),n.length>0&&(r+="?"+n.map((function(e){return e.key+"="+e.value})).join("&")),null!==(l=window.localStorage.getItem("jsonapiServerUri"))&&(r=l+r),c=P.getInstance(),e.next=10,c.getJson(r);case 10:return o=e.sent,e.abrupt("return",o.data);case 12:case"end":return e.stop()}}),e)})));return function(a,t){return e.apply(this,arguments)}}(),Re=function(e){var a=e.cloudContext,t=Object(n.useState)([]),r=Object(_.a)(t,2),l=r[0],c=r[1],o=Object(n.useState)([]),i=Object(_.a)(o,2),s=i[0],u=i[1],m=Object(n.useState)({key:"",direction:"ASC"}),d=Object(_.a)(m,2),p=d[0],b=d[1];return Object(n.useEffect)((function(){(function(){var e=Object(v.a)(j.a.mark((function e(){var t,n,r,l;return j.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return t=Te(a),n=t.map((function(e){return{key:e.name,label:e.labelName}})),c(n),e.next=5,J(t);case 5:return r=e.sent,e.next=8,De(a,p);case 8:l=e.sent,u(K(l,t,a,r));case 10:case"end":return e.stop()}}),e)})));return function(){return e.apply(this,arguments)}})()()}),[a,p]),Object(h.jsx)(le,{dataColumnList:l,dataRecordList:s,sortInfo:p,setSortInfo:b})},Ee=(t(55),function(e){var a=e.isActive,t=e.location,n=e.children;return Object(h.jsx)("li",{className:a?"active":"",children:Object(h.jsx)(i.b,{to:t,className:a?"is-active":"",children:n})})}),Ue=function(e){var a=e.tabColumnList;return Object(h.jsx)("nav",{className:"tabs",role:"navigation","aria-label":"Tabs",children:Object(h.jsx)("ul",{className:"tabs--primary nav nav-tabs",children:a.map((function(e,a){return Object(h.jsx)(Ee,{isActive:e.isActive,location:e.location,children:e.children},a)}))})})},ze=function(e){var a=e.menuType,t=e.menuName,r=Object(n.useContext)(A).cloudContext,l=Object(s.f)();Object(n.useEffect)((function(){if(r.cloudServiceProvider!==a)switch(r.cloudServiceProvider){case"aws_cloud":l.push(V(d[0]));break;case"k8s":l.push(V(p[0]))}}),[r]);var c=("aws_cloud"===a?d:p).map((function(e){return{isActive:e.labelName===t,location:V(e),children:e.labelName}}));return Object(h.jsx)(Ue,{tabColumnList:c})},Ve=function(){var e=L();return Object(h.jsx)(i.a,{basename:m,children:Object(h.jsxs)(s.c,{children:[Object(h.jsx)(s.a,{exact:!0,path:"/",children:Object(h.jsx)(C,{})}),Object(h.jsx)(s.a,{path:"/callback",children:Object(h.jsx)(O,{})}),Object(h.jsxs)(A.Provider,{value:e,children:[Object(h.jsxs)(s.a,{path:"/providers",children:[Object(h.jsx)(be,{}),Object(h.jsx)(Ae,{})]}),[].concat(Object(o.a)(d),Object(o.a)(p)).map((function(e){return Object(h.jsxs)(s.a,{path:V(e),children:[Object(h.jsx)(be,{}),Object(h.jsx)(ze,{menuType:e.cloudServiceProvider,menuName:e.labelName}),Object(h.jsx)(oe,{entityTypeId:M(e),column:e.entityColumn}),Object(h.jsx)("hr",{})]},M(e))})),e.cloudContextList.map((function(e){return Object(h.jsxs)(s.a,{path:G(e),children:[Object(h.jsx)(be,{}),Object(h.jsx)(Re,{cloudContext:e})]},e.labelName)}))]})]})})},Me=function(e){e&&e instanceof Function&&t.e(3).then(t.bind(null,62)).then((function(a){var t=a.getCLS,n=a.getFID,r=a.getFCP,l=a.getLCP,c=a.getTTFB;t(e),n(e),r(e),l(e),c(e)}))};c.a.render(Object(h.jsx)(r.a.StrictMode,{children:Object(h.jsx)(Ve,{})}),document.getElementById("root")),Me()}},[[56,1,2]]]);
-//# sourceMappingURL=main.ac7cbde5.chunk.js.map
+(this.webpackJsonpcloud_dashboard=this.webpackJsonpcloud_dashboard||[]).push([[0],{56:function(e,a,t){"use strict";t.r(a);var n=t(1),r=t.n(n),l=t(20),c=t.n(l),o=t(6),i=t(11),s=t(5),u="cloud_dashboard",m="/clouds/dashboard",d=[{cloudServiceProvider:"aws_cloud",labelName:"Instance",entityName:"instance",entityColumn:[{labelName:"Name",name:"name",type:"default"},{labelName:"Public IP",name:"public_ip",type:"default"},{labelName:"Instance State",name:"instance_state",type:"default"},{labelName:"Instance Type",name:"instance_type",type:"default"},{labelName:"Availability Zone",name:"availability_zone",type:"default"},{labelName:"Cost",name:"cost",type:"cost"},{labelName:"Created",name:"created",type:"datetime"}]},{cloudServiceProvider:"aws_cloud",labelName:"Image",entityName:"image",entityColumn:[{labelName:"Name",name:"name",type:"default"},{labelName:"AMI Name",name:"ami_name",type:"default"},{labelName:"AMI ID",name:"image_id",type:"default"},{labelName:"Status",name:"status",type:"default"},{labelName:"Source",name:"source",type:"default"},{labelName:"Root Device Type",name:"root_device_type",type:"default"},{labelName:"Visibility",name:"visibility",type:"boolean",value:["Public","Private"]},{labelName:"Created",name:"created",type:"datetime"}]},{cloudServiceProvider:"aws_cloud",labelName:"Security Group",entityName:"security_group",entityColumn:[{labelName:"Name",name:"name",type:"default"},{labelName:"Group ID",name:"group_id",type:"default"},{labelName:"VPC",name:"vpc_id",type:"join",info:{entityTypeId:"aws_cloud_vpc",keyColumn:"vpc_id",valueColumn:"name"}}]},{cloudServiceProvider:"aws_cloud",labelName:"Elastic IP",entityName:"elastic_ip",entityColumn:[{labelName:"Name",name:"name",type:"default"},{labelName:"Allocation ID",name:"allocation_id",type:"default"},{labelName:"Type",name:"elastic_ip_type",type:"default"},{labelName:"Elastic IP",name:"public_ip",type:"default"},{labelName:"Instance ID",name:"instance_id",type:"join",info:{entityTypeId:"aws_cloud_instance",keyColumn:"instance_id",valueColumn:"name"}},{labelName:"Private IP Address",name:"private_ip_address",type:"default"},{labelName:"Scope",name:"scope",type:"default"}]},{cloudServiceProvider:"aws_cloud",labelName:"Key Pair",entityName:"key_pair",entityColumn:[{labelName:"Key Pair Name",name:"key_pair_name",type:"default"},{labelName:"Key Fingerprint",name:"key_fingerprint",type:"default"}]},{cloudServiceProvider:"aws_cloud",labelName:"Volume",entityName:"volume",entityColumn:[{labelName:"Name",name:"name",type:"default"},{labelName:"Volume ID",name:"volume_id",type:"default"},{labelName:"IOPS",name:"iops",type:"default"},{labelName:"Size",name:"size",type:"default"},{labelName:"Availability Zone",name:"availability_zone",type:"default"},{labelName:"Volume Type",name:"volume_type",type:"default"},{labelName:"Attachment Information",name:"attachment_information",type:"join",info:{entityTypeId:"aws_cloud_instance",keyColumn:"instance_id",valueColumn:"name"}},{labelName:"State",name:"state",type:"default"},{labelName:"Created",name:"created",type:"datetime"}]},{cloudServiceProvider:"aws_cloud",labelName:"Snapshot",entityName:"snapshot",entityColumn:[{labelName:"Name",name:"name",type:"default"},{labelName:"Snapshot ID",name:"snapshot_id",type:"default"},{labelName:"Encrypted",name:"encrypted",type:"default"},{labelName:"Size",name:"size",type:"default"},{labelName:"Status",name:"status",type:"default"},{labelName:"Created",name:"created",type:"datetime"}]},{cloudServiceProvider:"aws_cloud",labelName:"Network Interface",entityName:"network_interface",entityColumn:[{labelName:"Name",name:"name",type:"default"},{labelName:"Subnet",name:"subnet_id",type:"join",info:{entityTypeId:"aws_cloud_subnet",keyColumn:"subnet_id",valueColumn:"name"}},{labelName:"VPC",name:"vpc_id",type:"join",info:{entityTypeId:"aws_cloud_vpc",keyColumn:"vpc_id",valueColumn:"name"}},{labelName:"Status",name:"status",type:"default"},{labelName:"Security Groups",name:"security_groups",type:"default"},{labelName:"Primary Private IP",name:"primary_private_ip",type:"default"},{labelName:"Secondary Private IPs",name:"secondary_private_ips",type:"default"}]},{cloudServiceProvider:"aws_cloud",labelName:"VPC",entityName:"vpc",entityColumn:[{labelName:"Name",name:"name",type:"default"},{labelName:"VPC",name:"vpc_id",type:"default"},{labelName:"State",name:"state",type:"default"},{labelName:"IPv4 CIDR",name:"cidr_block",type:"default"},{labelName:"IPv6 CIDR",name:"ipv6_cidr_blocks",type:"array"},{labelName:"Created",name:"created",type:"datetime"}]},{cloudServiceProvider:"aws_cloud",labelName:"Subnet",entityName:"subnet",entityColumn:[{labelName:"Name",name:"name",type:"default"},{labelName:"Subnet ID",name:"subnet_id",type:"default"},{labelName:"VPC",name:"vpc_id",type:"join",info:{entityTypeId:"aws_cloud_vpc",keyColumn:"vpc_id",valueColumn:"name"}},{labelName:"IPv4 CIDR",name:"cidr_block",type:"default"},{labelName:"State",name:"state",type:"default"},{labelName:"Region Name",name:"region_name",type:"default"},{labelName:"Zone Name",name:"zone_name",type:"default"},{labelName:"Network Border Group",name:"network_border_group",type:"default"},{labelName:"Zone Type",name:"zone_type",type:"default"},{labelName:"Parent Zone Type",name:"parent_zone_type",type:"default"},{labelName:"Created",name:"created",type:"datetime"}]},{cloudServiceProvider:"aws_cloud",labelName:"VPC Peering Connection",entityName:"vpc_peering_connection",entityColumn:[{labelName:"Name",name:"name",type:"default"},{labelName:"VPC Peering Connection ID",name:"vpc_peering_connection_id",type:"default"},{labelName:"Status",name:"status_code",type:"default"},{labelName:"Requester VPC",name:"requester_vpc_id",type:"join",info:{entityTypeId:"aws_cloud_vpc",keyColumn:"vpc_id",valueColumn:"name"}},{labelName:"Accepter VPC",name:"accepter_vpc_id",type:"join",info:{entityTypeId:"aws_cloud_vpc",keyColumn:"vpc_id",valueColumn:"name"}},{labelName:"Requester CIDR Blocks",name:"requester_cidr_block",type:"default"},{labelName:"Accepter CIDR Blocks",name:"accepter_cidr_block",type:"default"},{labelName:"Requester AWS Account",name:"requester_account_id",type:"default"},{labelName:"Accepter AWS Account",name:"accepter_account_id",type:"default"},{labelName:"Created",name:"created",type:"datetime"}]},{cloudServiceProvider:"aws_cloud",labelName:"Internet Gateway",entityName:"internet_gateway",entityColumn:[{labelName:"Name",name:"name",type:"default"},{labelName:"Internet Gateway ID",name:"internet_gateway_id",type:"default"},{labelName:"State",name:"state",type:"default"},{labelName:"VPC ID",name:"vpc_id",type:"default"},{labelName:"Created",name:"created",type:"default"}]},{cloudServiceProvider:"aws_cloud",labelName:"Career Gateway",entityName:"career_gateway",entityColumn:[{labelName:"Name",name:"name",type:"default"},{labelName:"Carrier Gateway ID",name:"carrier_gateway_id",type:"default"},{labelName:"State",name:"state",type:"default"},{labelName:"VPC ID",name:"vpc_id",type:"default"},{labelName:"Created",name:"created",type:"default"}]},{cloudServiceProvider:"aws_cloud",labelName:"Transit Gateway",entityName:"transit_gateway",entityColumn:[{labelName:"Name",name:"name",type:"default"},{labelName:"Transit Gateway ID",name:"transit_gateway_id",type:"default"},{labelName:"State",name:"state",type:"default"},{labelName:"Amazon Account ID",name:"account_id",type:"default"},{labelName:"Created",name:"created",type:"default"}]}],p=[{cloudServiceProvider:"k8s",labelName:"Node",entityName:"node",entityColumn:[{labelName:"Name",name:"name",type:"default"},{labelName:"State",name:"status",type:"default"},{labelName:"CPU (Request)",name:"cpu_request",type:"default"},{labelName:"CPU (Limit)",name:"cpu_limit",type:"default"},{labelName:"CPU (Usage)",name:"cpu_usage",type:"default"},{labelName:"Memory (Request)",name:"memory_request",type:"memory"},{labelName:"Memory (Limit)",name:"memory_limit",type:"memory"},{labelName:"Memory (Usage)",name:"memory_usage",type:"memory"},{labelName:"Pods (Applcation)",name:"pods_allocation",type:"default"},{labelName:"Created",name:"created",type:"datetime"}]},{cloudServiceProvider:"k8s",labelName:"Namespace",entityName:"namespace",entityColumn:[{labelName:"Name",name:"name",type:"default"},{labelName:"State",name:"status",type:"default"},{labelName:"Created",name:"created",type:"datetime"}]},{cloudServiceProvider:"k8s",labelName:"Deployment",entityName:"deployment",entityColumn:[{labelName:"Namespace",name:"namespace",type:"default"},{labelName:"Name",name:"name",type:"default"},{labelName:"Created",name:"created",type:"datetime"}]},{cloudServiceProvider:"k8s",labelName:"Pod",entityName:"pod",entityColumn:[{labelName:"Namespace",name:"namespace",type:"default"},{labelName:"Name",name:"name",type:"default"},{labelName:"Node",name:"node_name",type:"default"},{labelName:"State",name:"status",type:"default"},{labelName:"Restarts",name:"restarts",type:"default"},{labelName:"CPU (Usage)",name:"cpu_usage",type:"default"},{labelName:"Memory (Usage)",name:"memory_usage",type:"default"},{labelName:"Created",name:"created",type:"datetime"}]},{cloudServiceProvider:"k8s",labelName:"ReplicaSet",entityName:"replica_set",entityColumn:[{labelName:"Namespace",name:"namespace",type:"default"},{labelName:"Name",name:"name",type:"default"},{labelName:"Replica",name:"replicas",type:"default"},{labelName:"Created",name:"created",type:"datetime"}]},{cloudServiceProvider:"k8s",labelName:"CronJob",entityName:"cron_job",entityColumn:[{labelName:"Namespace",name:"namespace",type:"default"},{labelName:"Name",name:"name",type:"default"},{labelName:"Schedule",name:"schedule",type:"default"},{labelName:"Suspend",name:"suspend",type:"default"},{labelName:"Active",name:"active",type:"default"},{labelName:"Created",name:"created",type:"datetime"}]},{cloudServiceProvider:"k8s",labelName:"Job",entityName:"job",entityColumn:[{labelName:"Namespace",name:"namespace",type:"default"},{labelName:"Name",name:"name",type:"default"},{labelName:"Image",name:"image",type:"default"},{labelName:"Created",name:"created",type:"datetime"}]},{cloudServiceProvider:"k8s",labelName:"Service",entityName:"service",entityColumn:[{labelName:"Namespace",name:"namespace",type:"default"},{labelName:"Name",name:"name",type:"default"},{labelName:"Cluster IP",name:"cluster_ip",type:"default"},{labelName:"Created",name:"created",type:"datetime"}]},{cloudServiceProvider:"k8s",labelName:"Network Policy",entityName:"network_policy",entityColumn:[{labelName:"Namespace",name:"namespace",type:"default"},{labelName:"Name",name:"name",type:"default"},{labelName:"Created",name:"created",type:"datetime"}]},{cloudServiceProvider:"k8s",labelName:"Resource Quotas",entityName:"resource_quota",entityColumn:[{labelName:"Namespace",name:"namespace",type:"default"},{labelName:"Name",name:"name",type:"default"},{labelName:"Created",name:"created",type:"datetime"}]},{cloudServiceProvider:"k8s",labelName:"LimitRange",entityName:"limit_range",entityColumn:[{labelName:"Namespace",name:"namespace",type:"default"},{labelName:"Name",name:"name",type:"default"},{labelName:"Created",name:"created",type:"datetime"}]},{cloudServiceProvider:"k8s",labelName:"Priority Class",entityName:"priority_class",entityColumn:[{labelName:"Name",name:"name",type:"default"},{labelName:"Value",name:"value",type:"default"},{labelName:"Global Default",name:"global_default",type:"boolean",value:["TRUE","FALSE"]},{labelName:"Description",name:"description",type:"default"},{labelName:"Created",name:"created",type:"datetime"}]},{cloudServiceProvider:"k8s",labelName:"ConfigMap",entityName:"config_map",entityColumn:[{labelName:"Namespace",name:"namespace",type:"default"},{labelName:"Name",name:"name",type:"default"},{labelName:"Created",name:"created",type:"datetime"}]},{cloudServiceProvider:"k8s",labelName:"Secret",entityName:"secret",entityColumn:[{labelName:"Namespace",name:"namespace",type:"default"},{labelName:"Name",name:"name",type:"default"},{labelName:"Type",name:"secret_type",type:"default"},{labelName:"Created",name:"created",type:"datetime"}]},{cloudServiceProvider:"k8s",labelName:"Role",entityName:"role",entityColumn:[{labelName:"Namespace",name:"namespace",type:"default"},{labelName:"Name",name:"name",type:"default"},{labelName:"Created",name:"created",type:"datetime"}]},{cloudServiceProvider:"k8s",labelName:"Role Binding",entityName:"role_binding",entityColumn:[{labelName:"Namespace",name:"namespace",type:"default"},{labelName:"Name",name:"name",type:"default"},{labelName:"Role",name:"role_ref",type:"default"},{labelName:"Created",name:"created",type:"datetime"}]},{cloudServiceProvider:"k8s",labelName:"Cluster Role",entityName:"cluster_role",entityColumn:[{labelName:"Name",name:"name",type:"default"},{labelName:"Created",name:"created",type:"datetime"}]},{cloudServiceProvider:"k8s",labelName:"Cluster Role Binding",entityName:"cluster_role_binding",entityColumn:[{labelName:"Name",name:"name",type:"default"},{labelName:"Cluster Role",name:"role_ref",type:"default"},{labelName:"Created",name:"created",type:"datetime"}]},{cloudServiceProvider:"k8s",labelName:"Persistent Volume",entityName:"persistent_volume",entityColumn:[{labelName:"Name",name:"name",type:"default"},{labelName:"Capacity",name:"capacity",type:"default"},{labelName:"Access Modes",name:"access_modes",type:"default"},{labelName:"Reclaim Policy",name:"reclaim_policy",type:"default"},{labelName:"Status",name:"phase",type:"default"},{labelName:"Request",name:"claim_ref",type:"default"},{labelName:"Storage Class",name:"storage_class_name",type:"default"},{labelName:"Reason",name:"reason",type:"default"},{labelName:"Created",name:"created",type:"datetime"}]},{cloudServiceProvider:"k8s",labelName:"Persistent Volume Claim",entityName:"persistent_volume_claim",entityColumn:[{labelName:"Namespace",name:"namespace",type:"default"},{labelName:"Name",name:"name",type:"default"},{labelName:"Phase",name:"phase",type:"default"},{labelName:"VolumeName",name:"volume_name",type:"default"},{labelName:"Capacity",name:"capacity",type:"default"},{labelName:"Request",name:"request",type:"default"},{labelName:"Access Mode",name:"access_mode",type:"default"},{labelName:"Storage Class",name:"storage_class",type:"default"},{labelName:"Created",name:"created",type:"datetime"}]},{cloudServiceProvider:"k8s",labelName:"Storage Class",entityName:"storage_class",entityColumn:[{labelName:"Name",name:"name",type:"default"},{labelName:"Provisioner",name:"provisioner",type:"default"},{labelName:"Parameters",name:"parameters",type:"key-value"},{labelName:"Created",name:"created",type:"datetime"}]},{cloudServiceProvider:"k8s",labelName:"StatefulSet",entityName:"stateful_set",entityColumn:[{labelName:"Namespace",name:"namespace",type:"default"},{labelName:"Name",name:"name",type:"default"},{labelName:"Created",name:"created",type:"datetime"}]},{cloudServiceProvider:"k8s",labelName:"Ingress",entityName:"ingress",entityColumn:[{labelName:"Namespace",name:"namespace",type:"default"},{labelName:"Name",name:"name",type:"default"},{labelName:"Created",name:"created",type:"datetime"}]},{cloudServiceProvider:"k8s",labelName:"DaemonSet",entityName:"daemon_set",entityColumn:[{labelName:"Namespace",name:"namespace",type:"default"},{labelName:"Name",name:"name",type:"default"},{labelName:"Created",name:"created",type:"datetime"}]},{cloudServiceProvider:"k8s",labelName:"Endpoint",entityName:"endpoint",entityColumn:[{labelName:"Namespace",name:"namespace",type:"default"},{labelName:"Name",name:"name",type:"default"},{labelName:"Node",name:"node_name",type:"default"},{labelName:"Created",name:"created",type:"datetime"}]},{cloudServiceProvider:"k8s",labelName:"Event",entityName:"event",entityColumn:[{labelName:"Type",name:"k8s_event_type",type:"default"},{labelName:"Reason",name:"reason",type:"default"},{labelName:"Object Kind",name:"object_kind",type:"default"},{labelName:"Object Name",name:"object_name",type:"default"},{labelName:"Message",name:"message",type:"default"},{labelName:"Last Time Stamp",name:"time_stamp",type:"datetime"},{labelName:"Created",name:"created",type:"datetime"}]},{cloudServiceProvider:"k8s",labelName:"API Service",entityName:"api_service",entityColumn:[{labelName:"Name",name:"name",type:"default"},{labelName:"Created",name:"created",type:"datetime"}]},{cloudServiceProvider:"k8s",labelName:"ServiceAccount",entityName:"service_account",entityColumn:[{labelName:"Namespace",name:"namespace",type:"default"},{labelName:"Name",name:"name",type:"default"},{labelName:"Created",name:"created",type:"datetime"}]},{cloudServiceProvider:"k8s",labelName:"Horizontal Pod Autoscaler",entityName:"horizontal_pod_autoscaler",entityColumn:[{labelName:"Name",name:"name",type:"default"},{labelName:"Namespace",name:"namespace",type:"default"},{labelName:"Scale Target",name:"scale_target",type:"default"},{labelName:"Target CPU Utilization Percentage",name:"target_cpu_utilization_percentage",type:"default"}]},{cloudServiceProvider:"k8s",labelName:"Schedule",entityName:"schedule",entityColumn:[{labelName:"Name",name:"name",type:"default"},{labelName:"Kind",name:"kind",type:"default"},{labelName:"Namespace Name",name:"namespace_name",type:"default"},{labelName:"Resource Name",name:"resource_name",type:"default"},{labelName:"Launch Template Name",name:"launch_template_name",type:"default"},{labelName:"State",name:"state",type:"default"},{labelName:"Start",name:"start_time",type:"default"},{labelName:"Stop",name:"stop_time",type:"default"},{labelName:"Created",name:"created",type:"datetime"}]}],b=[{cloudServiceProvider:"aws_cloud",name:"ALL",labelName:"AWS Cloud (ALL)"},{cloudServiceProvider:"k8s",name:"ALL",labelName:"K8s (ALL)"}],f=[{labelName:"Name",name:"name",type:"default"},{labelName:"AMI Name",name:"field_image_id",type:"default"},{labelName:"Instance type",name:"field_instance_type",type:"default"},{labelName:"VPC",name:"field_vpc",type:"join",info:{entityTypeId:"aws_cloud_vpc",keyColumn:"vpc_id",valueColumn:"name"}},{labelName:"Max count",name:"field_max_count",type:"default"},{labelName:"Status",name:"field_workflow_status",type:"default"}],y=[{labelName:"Name",name:"name",type:"default"},{labelName:"Namespace",name:"field_namespace",type:"default"},{labelName:"Object",name:"field_object",type:"array"},{labelName:"Enable time scheduler",name:"field_enable_time_scheduler",type:"boolean",value:["On","Off"]},{labelName:"Workflow status",name:"field_workflow_status",type:"default"}],N=t(2),j=t.n(N),v=t(3),h=t(0),x=function(){var e=Object(v.a)(j.a.mark((function e(a,t,n){var r,l,c,o,i,s;return j.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return(r=new FormData).append("grant_type","authorization_code"),r.append("client_id",t),r.append("client_secret",u),r.append("code",a),r.append("redirect_uri",n),e.next=8,fetch("/oauth/token",{method:"POST",body:r});case 8:if((l=e.sent).ok){e.next=11;break}throw new Error("Authorization failed");case 11:return e.next=13,l.json();case 13:if(!("access_token"in(c=e.sent))){e.next=23;break}o=c.access_token,i=c.refresh_token,s=(new Date).getTime()+1e3*c.expires_in,window.localStorage.setItem("accessToken",o),window.localStorage.setItem("refreshToken",i),window.localStorage.setItem("expiresDatetime","".concat(s)),e.next=24;break;case 23:throw new Error("Authorization failed");case 24:case"end":return e.stop()}}),e)})));return function(a,t,n){return e.apply(this,arguments)}}(),g=function(){var e=Object(v.a)(j.a.mark((function e(){var a,t,n,r,l,c;return j.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return console.group("Authorization Code Grant"),e.next=3,fetch("/clouds/cloud_dashboard/config/client_id");case 3:return a=e.sent,e.next=6,fetch("/clouds/cloud_dashboard/config/callback_uri");case 6:if(t=e.sent,a.ok&&t.ok){e.next=13;break}return console.log("Client ID : No"),console.error("Authorization failed."),console.groupEnd(),window.location.href=m,e.abrupt("return");case 13:return e.next=15,a.json();case 15:return n=e.sent.id,e.next=18,t.json();case 18:if(r=e.sent.uri,console.log("Client ID : Yes"),l=new URL(window.location.href).searchParams,null!==(c=l.get("code"))){e.next=28;break}return console.log("Authorization code : No"),console.error("Authorization failed."),console.groupEnd(),window.location.href=m,e.abrupt("return");case 28:return console.log("authorization code : Yes"),e.prev=29,e.next=32,x(c,n,r);case 32:console.log("Access token : Yes"),console.groupEnd(),window.location.href=m+"/aws_cloud/instance",e.next=43;break;case 37:e.prev=37,e.t0=e.catch(29),console.log("Access token : No"),console.error("Authorization failed."),console.groupEnd(),window.location.href=m;case 43:case"end":return e.stop()}}),e,null,[[29,37]])})));return function(){return e.apply(this,arguments)}}(),O=function(){return Object(n.useEffect)((function(){g()}),[]),Object(h.jsx)("div",{className:"container-fluid",children:Object(h.jsx)("div",{className:"row",children:Object(h.jsx)("div",{className:"col",children:Object(h.jsx)("h2",{children:"Please waiting..."})})})})},_=t(4),C=function(){var e=Object(n.useState)(""),a=Object(_.a)(e,2),t=a[0],r=a[1],l=Object(n.useState)(""),c=Object(_.a)(l,2),o=c[0],i=c[1],s=function(){var e=Object(v.a)(j.a.mark((function e(){var a,t;return j.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return window.localStorage.removeItem("jsonapiServerUri"),e.next=3,fetch("/clouds/cloud_dashboard/config/client_id");case 3:return a=e.sent,e.next=6,fetch("/clouds/cloud_dashboard/config/callback_uri");case 6:if(t=e.sent,!a.ok||!t.ok){e.next=18;break}return e.t0=r,e.next=11,a.json();case 11:return e.t1=e.sent.id,(0,e.t0)(e.t1),e.t2=i,e.next=16,t.json();case 16:e.t3=e.sent.uri,(0,e.t2)(e.t3);case 18:case"end":return e.stop()}}),e)})));return function(){return e.apply(this,arguments)}}();return Object(n.useEffect)((function(){s()}),[]),Object(h.jsx)("div",{className:"container-fluid",children:Object(h.jsx)("div",{className:"row",children:Object(h.jsx)("div",{className:"col",children:Object(h.jsx)("form",{children:Object(h.jsx)("button",{type:"button",className:"btn btn-default",onClick:function(){var e="/oauth/authorize?response_type=code&client_id=".concat(t,"&redirect_uri=").concat(o);window.location.href=e},disabled:""===t,children:"login"})})})})})},k=t(31),w=t(32),S=function(){function e(){Object(k.a)(this,e),this.cache={}}return Object(w.a)(e,[{key:"getJson",value:function(){var e=Object(v.a)(j.a.mark((function e(a,t){var n,r,l,c;return j.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:if(!(a in this.cache)){e.next=5;break}if(n=this.cache[a],!((r=(new Date).getTime())>=n.unixtime&&r-n.unixtime<864e5)){e.next=5;break}return e.abrupt("return",n.response);case 5:return e.next=7,fetch(a,t);case 7:return e.next=9,e.sent.json();case 9:return l=e.sent,c=(new Date).getTime(),this.cache[a]={response:l,unixtime:c},e.abrupt("return",l);case 13:case"end":return e.stop()}}),e,this)})));return function(a,t){return e.apply(this,arguments)}}()}],[{key:"getInstance",value:function(){return this.instance}}]),e}();S.instance=new S;var P=S,I=t(61),L=function(){var e=Object(n.useState)(function(){var e=localStorage.getItem("cloudContext");return null!==e?JSON.parse(e):b[0]}()),a=Object(_.a)(e,2),t=a[0],r=a[1],l=Object(n.useState)(Object(o.a)(b)),c=Object(_.a)(l,2),i=c[0],s=c[1],u=Object(I.a)().i18n;Object(n.useEffect)((function(){var e=window.localStorage.getItem("languageName");null!==e&&u.changeLanguage(e)}),[]),Object(n.useEffect)((function(){(function(){var e=Object(v.a)(j.a.mark((function e(){var a,n,l,c,i;return j.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:a=["aws_cloud","k8s"],n=Object(o.a)(b),l=j.a.mark((function e(){var a,t;return j.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return a=i[c],e.next=3,P.getInstance().getJson("/jsonapi/cloud_config/".concat(a));case 3:t=e.sent.data.map((function(e){return{cloudServiceProvider:a,name:e.attributes.cloud_context,labelName:e.attributes.name}})),n=[].concat(Object(o.a)(n),Object(o.a)(t));case 5:case"end":return e.stop()}}),e)})),c=0,i=a;case 4:if(!(c<i.length)){e.next=9;break}return e.delegateYield(l(),"t0",6);case 6:c++,e.next=4;break;case 9:s(n),0===n.filter((function(e){return e.cloudServiceProvider===t.cloudServiceProvider&&e.name===t.name})).length&&r(b[0]);case 11:case"end":return e.stop()}}),e)})));return function(){return e.apply(this,arguments)}})()()}),[]);return{cloudContext:t,cloudContextList:i,dispatch:function(e){switch(e.type){case"setCloudContext":localStorage.setItem("cloudContext",JSON.stringify(e.message)),r(e.message);break;case"setLanguage":u.changeLanguage(e.message),window.localStorage.setItem("languageName",e.message)}}}},A=Object(n.createContext)({cloudContext:b[0],cloudContextList:[],dispatch:function(){}}),T=t(13),D=t(10),R=t.n(D),E=function(e,a){return"000000".concat(e).slice(-a)},U=function(e,a){for(var t=1,n=0;n<a;n+=1)t*=10;if(Math.floor(e)!==e)return"".concat(Math.round(e*t)/t);var r="".concat(e);if(0!==a){r+=".";for(var l=0;l<a;l+=1)r+="0"}return r},z=function(e,a,t){if(null===e)return"";switch(a.type){case"datetime":return function(e){var a=new Date(e),t=a.getFullYear(),n=a.getMonth()+1,r=a.getDate(),l=a.getHours(),c=a.getMinutes(),o="".concat(t,"/").concat(E(n,2),"/").concat(E(r,2));return o+" - ".concat(E(l,2),":").concat(E(c,2))}(e);case"memory":return e>=1073741824?"".concat(U(e/1073741824,2),"Gi"):e>=1048576?"".concat(U(e/1048576,2),"Mi"):e>=1024?"".concat(U(e/1024,2),"Ki"):U(e,2);case"key-value":var n,r=[],l=Object(T.a)(e);try{for(l.s();!(n=l.n()).done;){var c=n.value;r.push("".concat(c.item_key,":").concat(c.item_value))}}catch(u){l.e(u)}finally{l.f()}return r.join(", ");case"cost":return"$".concat(e);case"boolean":return e?a.value[0]:a.value[1];case"array":return e.map((function(e){return"".concat(e)})).join(", ");case"join":var o,i=Object(T.a)(t[a.info.entityTypeId]);try{for(i.s();!(o=i.n()).done;){var s=o.value;if(s.attributes[a.info.keyColumn]===e)return s.attributes[a.info.valueColumn]}}catch(u){i.e(u)}finally{i.f()}return"(".concat(e,")");default:return e}},V=function(e){return"/".concat(e.cloudServiceProvider,"/").concat(e.entityName)},M=function(e){return"".concat(e.cloudServiceProvider,"_").concat(e.entityName)},F=function(){var e=Object(v.a)(j.a.mark((function e(a,t){var n,r,l,c,o;return j.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:for(r in(n=[]).push({key:"page[limit]",value:"".concat(t.limit)}),n.push({key:"page[offset]",value:"".concat(t.offset)}),t.filter)n.push({key:"filter[".concat(r,"]"),value:t.filter[r]});return""!==t.sort.key&&n.push("ASC"===t.sort.direction?{key:"sort",value:t.sort.key}:{key:"sort",value:"-"+t.sort.key}),l="/jsonapi/".concat(a,"/").concat(a),n.length>0&&(l+="?"+n.map((function(e){return e.key+"="+e.value})).join("&")),null!==(c=window.localStorage.getItem("jsonapiServerUri"))&&(l=c+l),e.next=11,P.getInstance().getJson(l);case 11:return o=e.sent,e.abrupt("return",o.data);case 13:case"end":return e.stop()}}),e)})));return function(a,t){return e.apply(this,arguments)}}(),q=function(){var e=Object(v.a)(j.a.mark((function e(a){var t,n,r,l,c,i,s,u,m=arguments;return j.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:for(r in t=m.length>1&&void 0!==m[1]?m[1]:{},n=[],t)n.push({key:"filter[".concat(r,"]"),value:t[r]});l="/jsonapi/".concat(a,"/").concat(a),n.length>0&&(l+="?"+n.map((function(e){return e.key+"="+e.value})).join("&")),null!==(c=window.localStorage.getItem("jsonapiServerUri"))&&(l=c+l),i=[],s=P.getInstance();case 9:return e.next=12,s.getJson(l);case 12:if(u=e.sent,i=[].concat(Object(o.a)(i),Object(o.a)(u.data)),void 0===u.links.next){e.next=18;break}l=u.links.next.href,e.next=19;break;case 18:return e.abrupt("break",21);case 19:e.next=9;break;case 21:return e.abrupt("return",i);case 22:case"end":return e.stop()}}),e)})));return function(a){return e.apply(this,arguments)}}(),G=function(e){return"ALL"===e.name?"/".concat(e.cloudServiceProvider,"/server_template"):"/server_template/".concat(e.cloudServiceProvider,"/").concat(e.name)},J=function(){var e=Object(v.a)(j.a.mark((function e(a){var t,n,r,l,c;return j.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:t={},n=Object(T.a)(a),e.prev=2,n.s();case 4:if((r=n.n()).done){e.next=16;break}if("join"===(l=r.value).type){e.next=8;break}return e.abrupt("continue",14);case 8:if(!((c=l.info.entityTypeId)in t)){e.next=11;break}return e.abrupt("continue",14);case 11:return e.next=13,q(c);case 13:t[c]=e.sent;case 14:e.next=4;break;case 16:e.next=21;break;case 18:e.prev=18,e.t0=e.catch(2),n.e(e.t0);case 21:return e.prev=21,n.f(),e.finish(21);case 24:return e.abrupt("return",t);case 25:case"end":return e.stop()}}),e,null,[[2,18,21,24]])})));return function(a){return e.apply(this,arguments)}}(),K=function(e,a,t,n){var r,l=[],c=Object(T.a)(e);try{for(c.s();!(r=c.n()).done;){var o,i=r.value,s={id:i.id,value:{}},u=Object(T.a)(a);try{for(u.s();!(o=u.n()).done;){var m=o.value,d=i.attributes[m.name];s.value[m.name]="cloud_context"===m.name?t.labelName:z(d,m,n)}}catch(p){u.e(p)}finally{u.f()}l.push(s)}}catch(p){c.e(p)}finally{c.f()}return l},H=function(){var e=Object(v.a)(j.a.mark((function e(){var a,t,n;return j.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return a="/clouds/cloud_config_location",null!==(t=window.localStorage.getItem("jsonapiServerUri"))&&(a=t+a),n=P.getInstance(),e.next=6,n.getJson(a);case 6:return e.abrupt("return",e.sent);case 7:case"end":return e.stop()}}),e)})));return function(){return e.apply(this,arguments)}}(),Y=function(e,a){var t,n={},r=Object(T.a)(e);try{var l=function(){var e=t.value,r="".concat(e.Latitude,",").concat(e.Longitude);r in n?n[r].icon=new R.a.Icon({iconUrl:a}):n[r]={icon:new R.a.Icon({iconUrl:e.Items[0].Image}),cloudServiceProvider:e.Type,position:[parseFloat(e.Latitude),parseFloat(e.Longitude)],item:[]};var l=e.Items.map((function(a){return{iconUrl:a.Image,entityViewUrl:"aws_cloud"===e.Type?"/aws_cloud/instance":"/k8s/node",name:a.Name,positionLabel:"".concat(e.City,", ").concat(e.Country)}}));n[r].item=[].concat(Object(o.a)(n[r].item),Object(o.a)(l))};for(r.s();!(t=r.n()).done;)l()}catch(c){r.e(c)}finally{r.f()}return Object.values(n)},B=function(){var e=Object(v.a)(j.a.mark((function e(a,t){var n,r,l,c,o,i;return j.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return(n=new FormData).append("grant_type","refresh_token"),n.append("client_id",a),n.append("client_secret",u),n.append("refresh_token",t),e.next=7,fetch("/oauth/token",{method:"POST",body:n});case 7:if((r=e.sent).ok){e.next=10;break}throw new Error("Refresh failed");case 10:return e.next=12,r.json();case 12:if(!("access_token"in(l=e.sent))){e.next=22;break}c=l.access_token,o=l.refresh_token,i=(new Date).getTime()+1e3*l.expires_in,window.localStorage.setItem("accessToken",c),window.localStorage.setItem("refreshToken",o),window.localStorage.setItem("expiresDatetime","".concat(i)),e.next=23;break;case 22:throw new Error("Refresh failed");case 23:case"end":return e.stop()}}),e)})));return function(a,t){return e.apply(this,arguments)}}(),Z=function(){var e=Object(v.a)(j.a.mark((function e(){var a,t,n,r,l;return j.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:if(a=window.localStorage.getItem("accessToken"),t=window.localStorage.getItem("expiresDatetime"),null!==a&&null!=t){e.next=5;break}return window.location.href=m,e.abrupt("return");case 5:return console.group("Authorization Code Grant"),e.next=8,fetch("/clouds/cloud_dashboard/config/client_id");case 8:if(n=e.sent,r=window.localStorage.getItem("refreshToken"),n.ok&&null!==r){e.next=16;break}return console.log("Client ID : No"),console.error("Authorization failed."),console.groupEnd(),window.location.href=m,e.abrupt("return");case 16:return e.next=18,n.json();case 18:if(l=e.sent.id,console.log("Client ID : Yes"),!((new Date).getTime()<=parseInt(t,10))){e.next=25;break}return console.log("Token expired : No"),console.groupEnd(),e.abrupt("return");case 25:console.log("Token expired : Yes"),B(l,r).then((function(){console.log("Access token : Yes"),console.groupEnd()})).catch((function(){console.log("Access token : No"),console.error("Authorization failed."),console.groupEnd(),window.location.href=m}));case 27:case"end":return e.stop()}}),e)})));return function(){return e.apply(this,arguments)}}(),W=function(e){var a=e.value,t=e.setvalue,n=e.dataList,r=Object(I.a)().t;return Object(h.jsxs)("select",{className:"form-select form-control",value:a,onChange:function(e){t(e.currentTarget.value)},children:[Object(h.jsxs)("option",{value:"",children:["- ",r("All")," -"]}),n.map((function(e){return Object(h.jsx)("option",{value:e,children:e},e)}))]})},Q=function(){var e=Object(v.a)(j.a.mark((function e(a,t){var n,r,l;return j.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return n={},"ALL"!==a.name&&(n.cloud_context=a.name),r="".concat(a.cloudServiceProvider,"_").concat(t),e.next=5,q(r,n);case 5:return l=e.sent,e.abrupt("return",l.map((function(e){return e.attributes.name})));case 7:case"end":return e.stop()}}),e)})));return function(a,t){return e.apply(this,arguments)}}(),$=function(e){var a=e.columnKey,t=e.columnName,r=e.setColumnName,l=Object(n.useContext)(A).cloudContext,c=Object(n.useState)([]),o=Object(_.a)(c,2),i=o[0],s=o[1];return Object(n.useEffect)((function(){Q(l,a).then((function(e){s(e)}))}),[l]),Object(h.jsx)(W,{value:t,dataList:i,setvalue:function(e){r(e)}})},X=function(e){var a=e.children;return Object(h.jsx)("label",{className:"control-label",children:a})},ee=function(){var e=Object(v.a)(j.a.mark((function e(a,t,n,r){var l,c;return j.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return l="ALL"===a.name?"/cloud_dashboard/".concat(a.cloudServiceProvider,"/").concat(t,"/entity/count"):"/cloud_dashboard/".concat(a.cloudServiceProvider,"/").concat(a.name,"/").concat(t,"/entity/count"),""!==n&&(l+="?namespace=".concat(n)),""!==r&&(l+="?namespace_name=".concat(r)),null!==(c=window.localStorage.getItem("jsonapiServerUri"))&&(l=c+l),e.next=7,P.getInstance().getJson(l);case 7:return e.abrupt("return",e.sent.count);case 8:case"end":return e.stop()}}),e)})));return function(a,t,n,r){return e.apply(this,arguments)}}(),ae=function(e){var a=e.entityTypeId,t=e.namespace,r=e.namespaceName,l=e.itemCount,c=e.setItemCount,o=Object(n.useContext)(A).cloudContext,i=Object(I.a)().t;return Object(n.useEffect)((function(){ee(o,a,t,r).then((function(e){c(e)}))}),[o,t,r]),Object(h.jsxs)(X,{children:[i("ItemCount"),": ",l]})},te=function(e){var a=e.ariaLabel,t=e.onClick,n=e.children;return Object(h.jsx)("span",{"aria-label":a,onClick:t,children:Object(h.jsx)("span",{"aria-hidden":"true",children:n})})},ne=function(e){var a=e.pageIndex,t=e.setPageIndex,n=e.itemCount,r=Math.floor(1*(n+30-1)/30),l=0===a,c=0===n||a===r-1;return Object(h.jsx)("nav",{children:Object(h.jsxs)("ul",{className:"pagination",children:[Object(h.jsx)("li",{className:l?"disabled":"",children:Object(h.jsx)(te,{ariaLabel:"First",onClick:function(){t(0)},children:"\xab"})}),Object(h.jsx)("li",{className:l?"disabled":"",children:Object(h.jsx)(te,{ariaLabel:"Previous",onClick:function(){t(Math.max(0,a-1))},children:"\uff1c"})}),Object(h.jsx)("li",{children:Object(h.jsx)("span",{children:a+1})}),Object(h.jsx)("li",{className:c?"disabled":"",children:Object(h.jsx)(te,{ariaLabel:"Next",onClick:function(){t(Math.min(a+1,r-1))},children:"\uff1e"})}),Object(h.jsx)("li",{className:c?"disabled":"",children:Object(h.jsx)(te,{ariaLabel:"Last",onClick:function(){t(r-1)},children:"\xbb"})})]})})},re=function(e){var a=e.direction,t="ASC"===a?"glyphicon-chevron-up":"glyphicon-chevron-down",n="ASC"===a?"descending":"ascending";return Object(h.jsx)("span",{className:"icon glyphicon ".concat(t," icon-after"),"aria-hidden":"true","data-toggle":"tooltip","data-placement":"bottom",title:"","data-original-title":"Sort by ".concat(n," order")})},le=function(e){var a=e.dataColumn,t=e.sortInfo,n=e.setSortInfo,r=function(e){t.key===e?"ASC"!==t.direction?n({key:"",direction:"ASC"}):n({key:e,direction:"DESC"}):n({key:e,direction:"ASC"})};return t.key!==a.key?Object(h.jsx)("th",{onClick:function(){r(a.key)},children:a.label},a.key):Object(h.jsxs)("th",{className:"is-active",onClick:function(){r(a.key)},children:[a.label,Object(h.jsx)("a",{href:"#",onClick:function(e){e.preventDefault()},children:Object(h.jsx)(re,{direction:t.direction})})]},a.key)},ce=function(e){var a=e.dataRecord,t=e.dataColumnList,n=e.className;return Object(h.jsx)("tr",{className:n,children:t.map((function(e){return Object(h.jsx)("td",{children:e.key in a.value?a.value[e.key]:""},e.key)}))},a.id)},oe=function(e){var a=e.dataColumnList,t=e.dataRecordList,n=e.sortInfo,r=e.setSortInfo;return Object(h.jsx)("div",{className:"table-responsive",children:Object(h.jsxs)("table",{className:"responsive-enabled table table-hover table-striped",children:[Object(h.jsx)("thead",{children:Object(h.jsx)("tr",{children:a.map((function(e){return Object(h.jsx)(le,{dataColumn:e,sortInfo:n,setSortInfo:r})}))})}),Object(h.jsx)("tbody",{children:t.map((function(e,t){return Object(h.jsx)(ce,{dataRecord:e,dataColumnList:a,className:t%2===0?"odd":"even"})}))})]})})},ie=function(e){var a=e.entityTypeId,t=e.entityColumnList,r=e.namespace,l=e.namespaceName,c=e.pageIndex,o=Object(n.useContext)(A).cloudContext,i=Object(n.useState)([]),s=Object(_.a)(i,2),u=s[0],m=s[1],d=Object(n.useState)([]),p=Object(_.a)(d,2),b=p[0],f=p[1],y=Object(n.useState)({key:"",direction:"ASC"}),N=Object(_.a)(y,2),x=N[0],g=N[1];return Object(n.useEffect)((function(){(function(){var e=Object(v.a)(j.a.mark((function e(){var n,i,s,u,d;return j.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return n=t.map((function(e){return{key:e.name,label:e.labelName}})),m(n),e.next=4,J(t);case 4:return i=e.sent,s={},""!==r&&(s.namespace=r),""!==l&&(s.namespaceName=l),"ALL"!==o.name&&(s.cloud_context=o.name),u={limit:30,offset:30*c,filter:s,sort:x},e.next=12,F(a,u);case 12:d=e.sent,f(K(d,t,o,i));case 14:case"end":return e.stop()}}),e)})));return function(){return e.apply(this,arguments)}})()()}),[o,a,t,x,r,l,c]),Object(h.jsx)(oe,{dataColumnList:u,dataRecordList:b,sortInfo:x,setSortInfo:g})},se=function(e){var a=e.entityTypeId,t=e.column,r=Object(n.useState)(""),l=Object(_.a)(r,2),c=l[0],o=l[1],i=Object(n.useState)(""),s=Object(_.a)(i,2),u=s[0],m=s[1],d=Object(n.useState)(0),p=Object(_.a)(d,2),b=p[0],f=p[1],y=Object(n.useState)(0),N=Object(_.a)(y,2),j=N[0],v=N[1];Object(n.useEffect)((function(){""!==c&&m("")}),[c]),Object(n.useEffect)((function(){""!==u&&o("")}),[u]);var x=t.map((function(e){return e.labelName})).includes("Namespace"),g=t.map((function(e){return e.labelName})).includes("Namespace Name");return Object(h.jsxs)(h.Fragment,{children:[x?Object(h.jsxs)("div",{style:{marginTop:"2rem"},children:[Object(h.jsx)("label",{className:"control-label",children:"Namespace"}),Object(h.jsx)($,{columnKey:"namespace",columnName:c,setColumnName:o})]}):Object(h.jsx)(h.Fragment,{}),g?Object(h.jsxs)("div",{style:{marginTop:"2rem"},children:[Object(h.jsx)("label",{className:"control-label",children:"Namespace Name"}),Object(h.jsx)($,{columnKey:"namespace_name",columnName:u,setColumnName:m})]}):Object(h.jsx)(h.Fragment,{}),Object(h.jsx)("div",{style:{marginTop:"2rem"},children:Object(h.jsx)(ae,{entityTypeId:a,namespace:c,namespaceName:u,itemCount:b,setItemCount:f})}),Object(h.jsx)("div",{style:{marginTop:"2rem"},children:Object(h.jsx)(ie,{entityTypeId:a,entityColumnList:t,namespace:c,namespaceName:u,pageIndex:j})}),Object(h.jsx)(ne,{pageIndex:j,setPageIndex:v,itemCount:b})]})},ue=function(e){var a=e.to,t=e.onClick,n=e.className,r=e.children;return Object(h.jsx)("li",{className:n,children:Object(h.jsx)(i.b,{to:a,onClick:t,children:r})})},me=function(e){var a=e.href,t=e.className,n=e.children;return Object(h.jsx)("li",{className:t,children:Object(h.jsx)("a",{href:a,children:n})})},de=function(e){var a=e.menuName,t=e.linkPropsList;return Object(h.jsxs)("li",{className:"dropdown",children:[Object(h.jsxs)("a",{href:"#",className:"dropdown-toggle","data-toggle":"dropdown",children:["".concat(a)," ",Object(h.jsx)("span",{className:"caret"})]}),Object(h.jsx)("ul",{className:"dropdown-menu",role:"menu",children:t.map((function(e,a){return Object(h.jsx)(ue,{to:e.to,onClick:e.onClick,children:e.children},a)}))})]})},pe=function(e){var a=e.anchorPropsList;return Object(h.jsx)("ul",{className:"dropdown-menu",role:"menu",children:a.map((function(e,a){return Object(h.jsx)(me,{href:e.href,children:e.children},a)}))})},be=function(){var e=Object(n.useContext)(A),a=e.cloudContext,t=e.cloudContextList,r=e.dispatch,l=Object(I.a)().t;Object(n.useEffect)((function(){(function(){var e=Object(v.a)(j.a.mark((function e(){var a;return j.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return e.next=2,Z();case 2:if(null!==window.localStorage.getItem("jsonapiServerUri")){e.next=12;break}return e.next=5,fetch("/clouds/cloud_dashboard/config/jsonapi_server_uri");case 5:if(!(a=e.sent).ok){e.next=12;break}return e.t0=window.localStorage,e.next=10,a.json();case 10:e.t1=e.sent.uri,e.t0.setItem.call(e.t0,"jsonapiServerUri",e.t1);case 12:case"end":return e.stop()}}),e)})));return function(){return e.apply(this,arguments)}})()()}),[]);var c=function(e){r({type:"setCloudContext",message:e})};return Object(h.jsx)("header",{className:"navbar navbar-default navbar-fixed-top",role:"banner",children:Object(h.jsxs)("div",{className:"container-fluid",children:[Object(h.jsxs)("div",{className:"navbar-header",children:[Object(h.jsx)("div",{className:"region region-navigation",children:Object(h.jsx)("a",{className:"logo navbar-btn pull-left",href:"/",title:l("Home"),rel:"home",children:Object(h.jsx)("img",{src:"/themes/contrib/bootstrap_cloud/logo.svg",alt:l("Home")})})}),Object(h.jsxs)("button",{type:"button",className:"navbar-toggle","data-toggle":"collapse","data-target":"#navbar-collapse",children:[Object(h.jsx)("span",{className:"sr-only",children:"Toggle navigation"}),Object(h.jsx)("span",{className:"icon-bar"}),Object(h.jsx)("span",{className:"icon-bar"}),Object(h.jsx)("span",{className:"icon-bar"})]})]}),Object(h.jsx)("div",{className:"navbar-collapse collapse",children:Object(h.jsxs)("div",{className:"region region-navigation-collapsible",children:[Object(h.jsxs)("nav",{role:"navigation",className:"contextual-region",children:[Object(h.jsx)("h2",{className:"sr-only",children:"Main navigation"}),Object(h.jsxs)("ul",{className:"nav navbar-nav",role:"menu",children:[Object(h.jsx)(ue,{to:"/providers",children:l("Home")}),Object(h.jsx)(de,{menuName:a.labelName,linkPropsList:t.map((function(e){var a="aws_cloud"===e.cloudServiceProvider?d:p;return{to:V(a[0]),onClick:function(){return c(e)},children:e.labelName}}))}),Object(h.jsx)(de,{menuName:"Launch template",linkPropsList:t.map((function(e){return{to:G(e),onClick:function(){return c(e)},children:e.labelName}}))}),Object(h.jsxs)("li",{className:"dropdown",children:[Object(h.jsxs)("a",{href:"/clouds/design",className:"dropdown-toggle","data-toggle":"dropdown",children:["Design ",Object(h.jsx)("span",{className:"caret"})]}),Object(h.jsxs)("ul",{className:"dropdown-menu",role:"menu",children:[Object(h.jsxs)("li",{className:"dropdown-submenu",children:[Object(h.jsxs)("a",{href:"/clouds",className:"dropdown-submenu-toggle",children:["Projects ",Object(h.jsx)("span",{className:"caret"})]}),Object(h.jsx)(pe,{anchorPropsList:t.filter((function(e){return"k8s"===e.cloudServiceProvider&&"ALL"!==e.name})).map((function(e){return{href:"/clouds/design/project/".concat(e.name),children:e.labelName}}))})]}),Object(h.jsxs)("li",{className:"dropdown-submenu",children:[Object(h.jsxs)("a",{href:"/clouds",className:"dropdown-submenu-toggle",children:["Stores ",Object(h.jsx)("span",{className:"caret"})]}),Object(h.jsx)(pe,{anchorPropsList:[{href:"/clouds/design/store/k8s_cost_store",children:"K8s cost store"},{href:"/clouds/design/store/k8s_namespace_resource_store",children:"K8s namespace resource store"},{href:"/clouds/design/store/k8s_node_resource_store",children:"K8s node resource store"},{href:"/clouds/design/store/k8s_pod_resource_store",children:"K8s pod resource store"}]})]})]})]}),Object(h.jsx)(me,{href:"/admin/structure/cloud_config",children:l("CloudConfig")})]})]}),Object(h.jsxs)("nav",{role:"navigation",className:"contextual-region",children:[Object(h.jsx)("h2",{className:"sr-only",children:"User account menu"}),Object(h.jsxs)("ul",{className:"menu menu--account nav navbar-nav navbar-right",children:[Object(h.jsx)(me,{className:"first",href:"/user",children:l("User")}),Object(h.jsx)(ue,{className:"last",to:"/",onClick:function(){window.localStorage.removeItem("accessToken"),window.localStorage.removeItem("refreshToken"),window.localStorage.removeItem("expiresDatetime")},children:l("Logout")})]})]})]})})]})})},fe=t(26),ye=t(19),Ne=[{key:"en",value:"English"},{key:"ja",value:"\u65e5\u672c\u8a9e"}];fe.a.use(ye.e).init({resources:{en:{translation:{Language:"Language",CloudConfig:"Cloud Config",languageName:"English",Home:"Home",User:"User",Logout:"Logout",ItemCount:"Item Count",All:"All"}},ja:{translation:{Language:"\u4f7f\u7528\u8a00\u8a9e",CloudConfig:"\u7ba1\u7406",languageName:"\u65e5\u672c\u8a9e",Home:"\u30db\u30fc\u30e0",User:"\u30e6\u30fc\u30b6\u30fc\u60c5\u5831",Logout:"\u30ed\u30b0\u30a2\u30a6\u30c8",ItemCount:"\u4ef6\u6570",All:"\u3059\u3079\u3066"}}},lng:"en",interpolation:{escapeValue:!1}});fe.a;var je=t(60),ve=t(57),he=function(e){return e.map((function(e){return[e[1],e[0]]}))},xe={fillColor:"#ddc",fillOpacity:1,color:"#ddc",opacity:1,weight:1},ge=function(e){var a=e.coordinates;return Object(h.jsx)(h.Fragment,{children:a.map((function(e,a){return Object(h.jsx)(ve.a,{pathOptions:xe,positions:he(e)},a)}))})},Oe=function(e){var a=e.feature;return"Polygon"===a.geometry.type?Object(h.jsx)(ge,{coordinates:a.geometry.coordinates}):Object(h.jsx)(h.Fragment,{children:a.geometry.coordinates.map((function(e,a){return Object(h.jsx)(ge,{coordinates:e},a)}))})},_e=function(e){var a=e.mapData;return Object(h.jsx)(h.Fragment,{children:a.features.map((function(e,a){return Object(h.jsx)(Oe,{feature:e},a)}))})},Ce=t(58),ke=t(59),we=function(e){var a=e.cloudServiceProvider,t=e.label,r=Object(n.useContext)(A),l=r.cloudContextList,c=r.dispatch,o=function(e){var t=l.filter((function(t){return t.cloudServiceProvider===a&&t.labelName===e}));t.length>=1&&c({type:"setCloudContext",message:t[0]})};return Object(h.jsxs)("div",{children:[Object(h.jsx)("div",{children:Object(h.jsx)(i.b,{to:t.entityViewUrl,target:"_blank",rel:"noreferrer",onClick:function(){o(t.name)},children:Object(h.jsx)("img",{src:t.iconUrl,alt:t.name})})}),Object(h.jsxs)("div",{children:[Object(h.jsx)("strong",{children:Object(h.jsx)(i.b,{to:t.entityViewUrl,target:"_blank",rel:"noreferrer",onClick:function(){o(t.name)},children:t.name})}),Object(h.jsx)("br",{}),Object(h.jsxs)("span",{className:"location",children:[Object(h.jsx)("span",{className:"glyphicon glyphicon-map-marker"}),t.positionLabel]})]})]})},Se=function(e){var a=e.item;return Object(h.jsx)(Ce.a,{position:a.position,icon:a.icon,children:Object(h.jsx)(ke.a,{children:a.item.map((function(e,t){return Object(h.jsx)(we,{cloudServiceProvider:a.cloudServiceProvider,label:e},t)}))})})},Pe=function(){var e=Object(v.a)(j.a.mark((function e(){var a,t,n,r,l;return j.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return e.next=2,fetch("/clouds/cloud_dashboard/config/marker_icon_uri");case 2:if((a=e.sent).ok){e.next=5;break}return e.abrupt("return",[]);case 5:return e.next=7,a.json();case 7:return t=e.sent,n=t.uri,e.next=11,H();case 11:return r=e.sent,l=Y(r,n),e.abrupt("return",l);case 14:case"end":return e.stop()}}),e)})));return function(){return e.apply(this,arguments)}}(),Ie=function(){var e=Object(v.a)(j.a.mark((function e(){var a,t;return j.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return e.next=2,fetch("/clouds/cloud_dashboard/config/map_geojson_uri");case 2:if((a=e.sent).ok){e.next=5;break}return e.abrupt("return",{features:[]});case 5:return e.next=7,a.json();case 7:return t=e.sent,e.next=10,P.getInstance().getJson(t.uri);case 10:return e.abrupt("return",e.sent);case 11:case"end":return e.stop()}}),e)})));return function(){return e.apply(this,arguments)}}(),Le=function(){var e=Object(n.useState)([]),a=Object(_.a)(e,2),t=a[0],r=a[1],l=Object(n.useState)({features:[]}),c=Object(_.a)(l,2),o=c[0],i=c[1];return Object(n.useEffect)((function(){(function(){var e=Object(v.a)(j.a.mark((function e(){return j.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return e.t0=r,e.next=3,Pe();case 3:return e.t1=e.sent,(0,e.t0)(e.t1),e.t2=i,e.next=8,Ie();case 8:e.t3=e.sent,(0,e.t2)(e.t3);case 10:case"end":return e.stop()}}),e)})));return function(){return e.apply(this,arguments)}})()()}),[]),Object(h.jsxs)(je.a,{center:[0,0],zoom:2,scrollWheelZoom:!1,style:{height:500,backgroundColor:"#4e5d6c"},children:[Object(h.jsx)(_e,{mapData:o}),t.map((function(e,a){return Object(h.jsx)(Se,{item:e},a)}))]})},Ae=function(){var e=Object(n.useContext)(A),a=e.cloudContextList,t=e.dispatch,r=Object(I.a)(),l=r.t,c=r.i18n;return Object(h.jsxs)("div",{className:"container-fluid",children:[Object(h.jsx)("div",{className:"row",children:Object(h.jsx)("div",{className:"col",children:Object(h.jsx)(Le,{})})}),Object(h.jsx)("div",{className:"row",children:Object(h.jsx)("div",{className:"col",children:Object(h.jsxs)("form",{children:[Object(h.jsxs)("div",{className:"form-group",style:{marginTop:"2rem"},children:[Object(h.jsx)("label",{className:"control-label",children:l("Language")}),Object(h.jsx)("br",{}),Object(h.jsx)("div",{className:"btn-group",role:"group",children:Ne.map((function(e){return Object(h.jsx)("button",{type:"button",className:"btn btn-default",disabled:c.language===e.key,onClick:function(){t({type:"setLanguage",message:e.key})},children:e.value})}))})]}),Object(h.jsx)("div",{className:"form-group",style:{marginTop:"2rem"},children:Object(h.jsx)("label",{className:"control-label",children:"Cloud Service Providers"})}),Object(h.jsx)("div",{className:"table-responsive",style:{marginTop:"2rem"},children:Object(h.jsxs)("table",{className:"table table-hover table-striped sticky-enabled sticky-table",children:[Object(h.jsx)("thead",{children:Object(h.jsx)("th",{children:"Name"})}),Object(h.jsx)("tbody",{children:a.filter((function(e){return"ALL"!==e.name})).map((function(e,a){return Object(h.jsx)("tr",{children:Object(h.jsx)("td",{children:e.labelName})},a)}))})]})})]})})})]})},Te=function(e){switch(e.cloudServiceProvider){case"aws_cloud":return"ALL"!==e.name?f:[{labelName:"Cloud Service Provider ID",name:"cloud_context",type:"default"}].concat(Object(o.a)(f));case"k8s":return"ALL"!==e.name?y:[{labelName:"Cloud Service Provider ID",name:"cloud_context",type:"default"}].concat(Object(o.a)(y))}},De=function(){var e=Object(v.a)(j.a.mark((function e(a,t){var n,r,l,c,o;return j.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return n=[],"ALL"!==a.name&&n.push({key:"filter[cloud_context]",value:a.name}),""!==t.key&&n.push("ASC"===t.direction?{key:"sort",value:t.key}:{key:"sort",value:"-"+t.key}),r="/jsonapi/cloud_launch_template/".concat(a.cloudServiceProvider),n.length>0&&(r+="?"+n.map((function(e){return e.key+"="+e.value})).join("&")),null!==(l=window.localStorage.getItem("jsonapiServerUri"))&&(r=l+r),c=P.getInstance(),e.next=10,c.getJson(r);case 10:return o=e.sent,e.abrupt("return",o.data);case 12:case"end":return e.stop()}}),e)})));return function(a,t){return e.apply(this,arguments)}}(),Re=function(e){var a=e.cloudContext,t=Object(n.useState)([]),r=Object(_.a)(t,2),l=r[0],c=r[1],o=Object(n.useState)([]),i=Object(_.a)(o,2),s=i[0],u=i[1],m=Object(n.useState)({key:"",direction:"ASC"}),d=Object(_.a)(m,2),p=d[0],b=d[1];return Object(n.useEffect)((function(){(function(){var e=Object(v.a)(j.a.mark((function e(){var t,n,r,l;return j.a.wrap((function(e){for(;;)switch(e.prev=e.next){case 0:return t=Te(a),n=t.map((function(e){return{key:e.name,label:e.labelName}})),c(n),e.next=5,J(t);case 5:return r=e.sent,e.next=8,De(a,p);case 8:l=e.sent,u(K(l,t,a,r));case 10:case"end":return e.stop()}}),e)})));return function(){return e.apply(this,arguments)}})()()}),[a,p]),Object(h.jsx)(oe,{dataColumnList:l,dataRecordList:s,sortInfo:p,setSortInfo:b})},Ee=(t(55),function(e){var a=e.isActive,t=e.location,n=e.children;return Object(h.jsx)("li",{className:a?"active":"",children:Object(h.jsx)(i.b,{to:t,className:a?"is-active":"",children:n})})}),Ue=function(e){var a=e.tabColumnList;return Object(h.jsx)("nav",{className:"tabs",role:"navigation","aria-label":"Tabs",children:Object(h.jsx)("ul",{className:"tabs--primary nav nav-tabs",children:a.map((function(e,a){return Object(h.jsx)(Ee,{isActive:e.isActive,location:e.location,children:e.children},a)}))})})},ze=function(e){var a=e.menuType,t=e.menuName,r=Object(n.useContext)(A).cloudContext,l=Object(s.f)();Object(n.useEffect)((function(){if(r.cloudServiceProvider!==a)switch(r.cloudServiceProvider){case"aws_cloud":l.push(V(d[0]));break;case"k8s":l.push(V(p[0]))}}),[r]);var c=("aws_cloud"===a?d:p).map((function(e){return{isActive:e.labelName===t,location:V(e),children:e.labelName}}));return Object(h.jsx)(Ue,{tabColumnList:c})},Ve=function(){var e=L();return Object(h.jsx)(i.a,{basename:m,children:Object(h.jsxs)(s.c,{children:[Object(h.jsx)(s.a,{exact:!0,path:"/",children:Object(h.jsx)(C,{})}),Object(h.jsx)(s.a,{path:"/callback",children:Object(h.jsx)(O,{})}),Object(h.jsxs)(A.Provider,{value:e,children:[Object(h.jsxs)(s.a,{path:"/providers",children:[Object(h.jsx)(be,{}),Object(h.jsx)(Ae,{})]}),[].concat(Object(o.a)(d),Object(o.a)(p)).map((function(e){return Object(h.jsxs)(s.a,{path:V(e),children:[Object(h.jsx)(be,{}),Object(h.jsx)(ze,{menuType:e.cloudServiceProvider,menuName:e.labelName}),Object(h.jsx)(se,{entityTypeId:M(e),column:e.entityColumn}),Object(h.jsx)("hr",{})]},M(e))})),e.cloudContextList.map((function(e){return Object(h.jsxs)(s.a,{path:G(e),children:[Object(h.jsx)(be,{}),Object(h.jsx)(Re,{cloudContext:e})]},e.labelName)}))]})]})})},Me=function(e){e&&e instanceof Function&&t.e(3).then(t.bind(null,62)).then((function(a){var t=a.getCLS,n=a.getFID,r=a.getFCP,l=a.getLCP,c=a.getTTFB;t(e),n(e),r(e),l(e),c(e)}))};c.a.render(Object(h.jsx)(r.a.StrictMode,{children:Object(h.jsx)(Ve,{})}),document.getElementById("root")),Me()}},[[56,1,2]]]);
+//# sourceMappingURL=main.4a7f9488.chunk.js.map
